% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{abc_accept}
\alias{abc_accept}
\title{Return result for ABC whether the particles match the data}
\usage{
abc_accept(accept, epsilon)
}
\arguments{
\item{accept}{A logical vector with one value for each particle in
the simulated model.}

\item{epsilon}{A numeric vector with the tolerance for the
generation.}
}
\value{
\code{list(accept = accept, epsilon = epsilon)}
}
\description{
Return result for ABC whether the particles match the data
}
