% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{as.data.frame.SimInf_abc_smc}
\alias{as.data.frame.SimInf_abc_smc}
\title{Coerce to data frame}
\usage{
\method{as.data.frame}{SimInf_abc_smc}(x, ...)
}
\arguments{
\item{x}{any \R object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
Coerce to data frame
}
