% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_model.R
\docType{methods}
\name{boxplot-methods}
\alias{boxplot-methods}
\alias{boxplot}
\alias{boxplot,SimInf_model-method}
\title{Box Plots}
\usage{
\S4method{boxplot}{SimInf_model}(x, ...)
}
\arguments{
\item{x}{The \code{model} to plot}

\item{...}{Additional arguments affecting the plot produced.}
}
\description{
Produce box-and-whisker plot(s) of the number of individuals in
each compartment.
}
