% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{boxplot,SimInf_model-method}
\alias{boxplot,SimInf_model-method}
\title{Box plot of number of individuals in each compartment}
\usage{
\S4method{boxplot}{SimInf_model}(x, compartments = NULL, index = NULL, ...)
}
\arguments{
\item{x}{The \code{SimInf_model} object containing the trajectory
data.}

\item{compartments}{Specify the names of the compartments to
include.  Can be a character vector (e.g., \code{c("S", "I",
"R")}) or a formula (e.g., \code{~S + I + R}). If \code{NULL}
(default), all compartments in the model are included.}

\item{index}{Indices of the nodes to include in the plot.  If
\code{NULL} (default), data from all nodes are pooled
together.  If a vector (e.g., \code{1:2}), only data from
those specific nodes are used.}

\item{...}{Additional graphical arguments passed to
\code{boxplot()}.}
}
\description{
Produce box-and-whisker plots summarizing the distribution of the
number of individuals in specified model compartments. The plots
aggregate data across all time points in \code{tspan} and the
selected nodes (specified by \code{index}).
}
\details{
This function is useful for visualizing the variability and range
of compartment counts across the entire simulation trajectory.
}
\examples{
## For reproducibility, specify the number of threads.
set_num_threads(1)

## Create an 'SIR' model with 10 nodes.
model <- SIR(
  u0 = data.frame(
    S = rep(99, 10),
    I = rep(1, 10),
    R = rep(0, 10)
  ),
  tspan = 1:100,
  beta = 0.16,
  gamma = 0.077
)

## Run the model with a fixed seed for reproducibility.
result <- run(model, seed = 22)

## Create a boxplot for all compartments across all nodes and time
## points.
boxplot(result)

## Create a boxplot for the S and I compartments, but only for
## nodes 1 and 2.
boxplot(result, compartments = c("S", "I"), index = 1:2)
}
