% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_end_t_arg}
\alias{check_end_t_arg}
\title{Check arguments for interval endpoints}
\usage{
check_end_t_arg(len, ...)
}
\arguments{
\item{len}{Exprected length of each of the interval endpoint
vectors}

\item{...}{The arguments to check}
}
\value{
invisible(NULL)
}
\description{
Raise an error if any of the arguments are not ok.
}
\keyword{internal}

