% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_gdata_arg}
\alias{check_gdata_arg}
\title{Check arguments for 'gdata'}
\usage{
check_gdata_arg(...)
}
\arguments{
\item{...}{The arguments to check}
}
\value{
invisible(NULL)
}
\description{
Raise an error if any of the arguments are not ok.
}
\keyword{internal}

