% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_infectious_pressure_arg}
\alias{check_infectious_pressure_arg}
\title{Check integer arguments}
\usage{
check_infectious_pressure_arg(len, ...)
}
\arguments{
\item{len}{Expected length of the infectious pressure vector}

\item{...}{The arguments to check}
}
\value{
invisible(NULL)
}
\description{
Raise an error if any of the arguments are non-integer.
}
\keyword{internal}

