% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_integer_arg}
\alias{check_integer_arg}
\title{Check integer arguments}
\usage{
check_integer_arg(...)
}
\arguments{
\item{...}{The arguments to check}
}
\value{
invisible(NULL)
}
\description{
Raise an error if any of the arguments are non-integer.
}
\keyword{internal}
