% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_null_arg}
\alias{check_null_arg}
\title{Check for null arguments}
\usage{
check_null_arg(...)
}
\arguments{
\item{...}{The argumens to check}
}
\value{
invisible(NULL)
}
\description{
Raise an error if any of the arguments are null.
}
\keyword{internal}

