% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_numeric_arg}
\alias{check_numeric_arg}
\title{Check numeric arguments}
\usage{
check_numeric_arg(...)
}
\arguments{
\item{...}{The argumens to check}
}
\value{
invisible(NULL)
}
\description{
Raise an error if any of the arguments are non-numeric.
}
\keyword{internal}

