% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/run.r
\name{check_strategy}
\alias{check_strategy}
\title{Check strategy argument}
\usage{
check_strategy(threads, strategy)
}
\arguments{
\item{threads}{Number of threads.}

\item{strategy}{The parallelization strategy.}
}
\value{
The strategy
}
\description{
Check strategy argument
}
\keyword{internal}

