% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/run.r
\name{check_threads}
\alias{check_threads}
\title{Check threads argument}
\usage{
check_threads(threads)
}
\arguments{
\item{threads}{Number of threads.}
}
\value{
Integer scalar with number of threads.
}
\description{
Check threads argument
}
\keyword{internal}

