% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/external_events.r
\docType{methods}
\name{coerce-external_events-list-method}
\alias{coerce,external_events,list-method}
\alias{coerce-external_events-list-method}
\title{Coerce S4 class \code{external_events} to a \code{list}}
\arguments{
\item{from}{The external_events \code{object}}
}
\value{
\code{list}
}
\description{
Coerce S4 class \code{external_events} to a \code{list}
}
\keyword{methods}

