% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/external_events.r
\docType{methods}
\name{coerce-list-external_events-method}
\alias{coerce,list,external_events-method}
\alias{coerce-list-external_events-method}
\title{Coerce \code{list} to a S4 class \code{external_events}}
\arguments{
\item{from}{The list \code{object}}
}
\value{
S4 class \code{external_events}
}
\description{
Coerce \code{list} to a S4 class \code{external_events}
}
\keyword{methods}

