% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SISe.r
\docType{methods}
\name{coerce-siminf_model-SISe-method}
\alias{coerce,siminf_model,SISe-method}
\alias{coerce-siminf_model-SISe-method}
\title{Coerce a \code{siminf_model} to a \code{SISe} model}
\arguments{
\item{from}{The siminf_model \code{object}}
}
\value{
\code{SISe} model
}
\description{
Coerce a \code{siminf_model} to a \code{SISe} model
}
\keyword{methods}

