% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SISe3.r
\docType{methods}
\name{coerce-siminf_model-SISe3-method}
\alias{coerce,siminf_model,SISe3-method}
\alias{coerce-siminf_model-SISe3-method}
\title{Coerce a \code{siminf_model} to a \code{SISe3} model}
\arguments{
\item{from}{The siminf_model \code{object}}
}
\value{
\code{SISe3} model
}
\description{
Coerce a \code{siminf_model} to a \code{SISe3} model
}
\keyword{methods}

