% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_mparse.R
\name{compile_mparse}
\alias{compile_mparse}
\title{Function to compile custom \code{\link{SimInf_model}}.}
\usage{
compile_mparse(model, name)
}
\description{
This function compiles a model specified using \code{\link{mparse}}
function, and produces a customised named \code{\link{SimInf_model}}
object that can be run in the usual way. This is useful for routines
that require multiple calls to the \code{run} method for 
\code{\link{SimInf_model}} objects created using \code{\link{mparse}},
which avoids the need to re-compile the model each time \code{run} is
called.
}
\examples{
## Create an SIR model object using \\code{mparse}.
transitions <- c("S -> beta*S*I -> I", "I -> gamma*I -> R")
compartments <- c("S", "I", "R")
u0 <- data.frame(S = 99, I = 1, R = 0)
model <- mparse(transitions = transitions, compartments = compartments,
    gdata = c(beta = 0.16, gamma = 0.077), u0 = u0, tspan = 1:100)

## Run the SIR model and plot the result.
set.seed(22)
result <- run(model)
plot(result)

## compile the model first and then re-run
## and plot the result
set.seed(22)
model <- compile_mparse(model, "SIR")
result <- run(model)
plot(result)
}
