% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{continue_abc}
\alias{continue_abc}
\alias{continue_abc,SimInf_abc-method}
\title{Run more generations of ABC SMC}
\usage{
continue_abc(object, ...)

\S4method{continue_abc}{SimInf_abc}(
  object,
  tolerance = NULL,
  ...,
  verbose = getOption("verbose", FALSE),
  post_gen = NULL,
  init_model = NULL
)
}
\arguments{
\item{object}{The \code{SimInf_abc} object to continue from.}

\item{...}{Further arguments to be passed to the
\code{SimInf_abc@fn}.}

\item{tolerance}{A numeric matrix (number of summary statistics
\eqn{\times} number of generations) where each column contains
the tolerances for a generation and each row contains a
sequence of gradually decreasing tolerances. Can also be a
numeric vector if there is only one summary statistic. The
tolerance determines the number of generations of ABC-SMC to
run.}

\item{verbose}{prints diagnostic messages when \code{TRUE}. The
default is to retrieve the global option \code{verbose} and
use \code{FALSE} if it is not set.}

\item{post_gen}{An optional function that, if non-NULL, is applied
after each completed generation. The function must accept one
argument of type \code{SimInf_abc} with the current state of
the fitting process. This function can be useful to, for
example, save and inspect intermediate results.}

\item{init_model}{An optional function that, if non-NULL, is
applied before running each proposal. The function must accept
one argument of type \code{SimInf_model} with the current
model of the fitting process. This function can be useful to
specify the initial state of \code{u0} or \code{v0} of the
model before running a trajectory with proposed parameters.}
}
\value{
A \code{SimInf_abc} object.
}
\description{
Run more generations of ABC SMC
}
