% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/external_events.r
\docType{class}
\name{external_events-class}
\alias{external_events-class}
\title{Class \code{"external_events"}}
\description{
Class to handle the external events
}
\section{Slots}{

\describe{
  \item{E}{
    Sparse matrix (\eqn{Ncompartments \times (4 * Nobs)}) of
    object class \code{"\linkS4class{dgCMatrix}"}. Each event
    type contains the number of observable states (Nobs) columns.
    Currently 4 types of events are implemented, see \code{event}.
    For the events EXIT_EVENT and EXTERNAL_TRANSFER_EVENT, a non-zero entry
    in element \code{i} of select column \code{j} indicate the
    compartments to sample individuals from.
    For the event ENTER_EVENT, all individuals enter first non-zero
    compartment, i.e. a non-zero entry in element \code{i} of select
    column \code{j}.
    For the INTERNAL_TRANSFER_EVENT, a non-zero entry
    in element \code{i} of select column \code{j} indicate the
    compartments to sample individuals from. The value of element
    \code{i} of select column \code{j} determines the target
    compartment. The target compartment for sampled individuals is
    given by adding the value.
  }
  \item{event}{
    Integer vector of length \code{len} with external events.
    The following four events are implemented; EXIT_EVENT = 0,
    ENTER_EVENT = 1, INTERNAL_TRANSFER_EVENT = 2, and
    EXTERNAL_TRANSFER_EVENT = 3.
  }
  \item{time}{
    Integer vector of length \code{len} with the time for external event.
  }
  \item{select}{
    Integer vector of length \code{len}. Column j in the event matrix
    that determines the hidden states to sample from.
  }
  \item{node}{
    Integer vector of length \code{len}. The source herd of the event i.
  }
  \item{dest}{
    Integer vector of length \code{len}. The dest herd of the event i.
  }
  \item{n}{
    Integer vector of length \code{len}. The number of individuals in the
    external event. n[i] >= 0.
  }
  \item{proportion}{
    Numeric vector of length \code{len}. If n[i] equals zero,
    then the number of individuals to sample is calculated by summing
    the number of individuals in the hidden states determined by
    select[i] and multiplying with the proportion.
    0 <= proportion[i] <= 1.
  }
  \item{len}{
    Number of scheduled external events.
  }
}
}
\keyword{classes}
\keyword{methods}

