% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/external_events.r
\name{external_events}
\alias{external_events}
\title{Create S4 class \code{external_events}}
\usage{
external_events(E = NULL, events = NULL)
}
\arguments{
\item{E}{Sparse matrix (\eqn{Ncompartments \times (4 * Nselect)}) of
object class \code{"\linkS4class{dgCMatrix}"}.}

\item{events}{A \code{data.frame} with events.}
}
\value{
S4 class \code{external_events}
}
\description{
The argument events must be a \code{data.frame} with the following
columns:
\itemize{
  \item{event}{
    The event type
  }
  \item{time}{
    The time of the event
  }
  \item{select}{
    The internal category of the event
  }
  \item{node}{
    The herd of the event
  }
  \item{dest}{
    The destination herd of an animal movement
  }
  \item{n}{
    The number of individuals
  }
  \item{prop}{
    The proportion of individuals. If n[i] equals zero,
    then the number of individuals to sample is calculated by summing
    the number of individuals in the hidden states determined by
    select[i] and multiplying with the proportion. 0 <= p[i] <= 1.
  }
  \item{thread}{
    Optional column with thread id of the event.
  }
}
}

