% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_individual_events.R
\name{get_individuals}
\alias{get_individuals}
\alias{get_individuals,SimInf_individual_events-method}
\title{Extract individuals from \code{SimInf_individual_events}}
\usage{
get_individuals(x, time = NULL)

\S4method{get_individuals}{SimInf_individual_events}(x, time = NULL)
}
\arguments{
\item{x}{an individual events object of class
\code{SimInf_individual_events}.}

\item{time}{the time-point for the lookup of individuals. Default
is \code{NULL} which means to extract the individuals at the
minimum time-point in the events object \code{x}.}
}
\value{
a \code{data.frame} with the columns \code{id},
    \code{node}, and \code{age}.
}
\description{
Lookup individuals, in which node they are located and their age
at a specified time-point.
}
