% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterated-filtering.R
\name{if2}
\alias{if2}
\alias{if2,SimInf_model-method}
\title{Iterated filtering algorithm}
\usage{
if2(model, verbose = getOption("verbose", FALSE))

\S4method{if2}{SimInf_model}(model, verbose = getOption("verbose", FALSE))
}
\arguments{
\item{model}{The model to simulate data from.}
}
\description{
Iterated filtering algorithm
}
\references{
\Ionides2015
}
