% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_indiv_events.R
\name{individuals}
\alias{individuals}
\alias{individuals,SimInf_indiv_events-method}
\title{Extract individuals from \code{SimInf_indiv_events}}
\usage{
individuals(x, at = NULL)

\S4method{individuals}{SimInf_indiv_events}(x, at = NULL)
}
\arguments{
\item{x}{an individual events object of class
\code{SimInf_indiv_events}.}

\item{at}{the time-point for the lookup of individuals. Default is
\code{NULL} which means to extract the individuals at the
minimum time-point in the events object \code{x}.}
}
\value{
a \code{data.frame} with the columns \code{id},
    \code{node}, and \code{age}.
}
\description{
Lookup individuals, in which node they are located and their age
at a specified time-point.
}
