% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SISe.R, R/SISe3.R
\docType{methods}
\name{infected}
\alias{infected}
\alias{infected,SISe-method}
\alias{infected,SISe3-method}
\title{Infected}
\usage{
infected(model, ...)

\S4method{infected}{SISe}(model, i = NULL, by = 1, ...)

\S4method{infected}{SISe3}(model, age = 1:3, i = NULL, by = 1, ...)
}
\arguments{
\item{model}{The \code{model} to extract the infected from}

\item{...}{Additional arguments affecting the measure}

\item{i}{Indices specifying the nodes to include when extracting
the number of infected. Default is NULL, which includes all nodes.}

\item{by}{The number to increment the sequence of time points
starting from 1. Default is 1, which gives the number of
infected at every time point.}

\item{age}{For models with age categories, the age category to
extract.}
}
\description{
Extracts the number of infected
}
\keyword{methods}

