% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/model_builder.R
\docType{methods}
\name{init}
\alias{init}
\alias{init,SimInf_mparse-method}
\title{Init a \code{SimInf_mparse} object}
\usage{
init(model, u0 = NULL, tspan = NULL)

\S4method{init}{SimInf_mparse}(model, u0 = NULL, tspan = NULL)
}
\arguments{
\item{model}{The \code{SimInf_mparse} object to initialize.}

\item{u0}{A \code{data.frame} (or an object that can be coerced to
a \code{data.frame}) with the initial state in each node.}

\item{tspan}{A vector (length >= 2) of increasing time points
where the state of each node is to be returned. Can be either
an \code{integer} or a \code{Date} vector. A \code{Date}
vector is coerced to a numeric vector as days, where
\code{tspan[1]} becomes the day of the year of the first year
of \code{tspan}. The dates are added as names to the numeric
vector.}
}
\value{
a \code{SimInf_model} object
}
\description{
Init a \code{SimInf_mparse} object
}
