% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambert.R
\name{lambertW0}
\alias{lambertW0}
\title{Lambert W0 function}
\usage{
lambertW0(x)
}
\arguments{
\item{x}{A numeric vector of values.}
}
\value{
A numeric vector of the same length as \code{x} containing
    the principal branch of the Lambert W function evaluated at
    each element.
}
\description{
The principal branch of the Lambert W function, which solves
\eqn{W e^W = x}{W * exp(W) = x}. Defined for \eqn{x >= -1/e}, with
output in the range \eqn{[-1, \infty)}{[-1, Inf)}. The value is
calculated using the GNU Scientific Library (GSL).
}
\examples{
## Should equal 1, as 1 * exp(1) = e.
lambertW0(exp(1))

## Should equal 0, as 0 * exp(0) = 0.
lambertW0(0)

## Should equal -1.
lambertW0(-exp(-1))
}
\references{
GNU Scientific Library <https://www.gnu.org/software/gsl/>
}
