% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{ldata_add_spatial_coupling}
\alias{ldata_add_spatial_coupling}
\title{Add information about spatial coupling between nodes to 'ldata'}
\usage{
ldata_add_spatial_coupling(ldata, distance, metric = 1)
}
\arguments{
\item{ldata}{Matrix with local model parameters for each node.}

\item{distance}{Sparse matrix
(\code{\link[Matrix:dgCMatrix-class]{dgCMatrix}}) with
distances between nodes, see \code{\link{distance_matrix}}.}

\item{metric}{The type of value to calculate for each neighbor: 0)
degree, 1) distance, and 2) 1 / (distance * distance). Default
is \code{1}.}
}
\value{
matrix
}
\description{
A utility function to combine local model parameters
(\code{ldata}) and spatial coupling to other nodes and add the
result to \code{ldata}.
}
\details{
Format for ldata: the first n indicies (1, 2, ..., n) are the
local model parameters, i.e, the indata to the function. They are
followed by the neighbor data, pairs of (index, value) and then a
stop pair (-1, 0) where 'index' is the zero-based index to the
neighbor and the value is determined by the metric argument.
}
\examples{
d <- distance_matrix(nodes$x, nodes$y, cutoff = 5000)
ldata <- rbind(nodes$x, nodes$y)
ldata <- ldata_add_spatial_coupling(ldata, d)
}
