% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcmc.R
\name{length,SimInf_pmcmc-method}
\alias{length,SimInf_pmcmc-method}
\title{Length of the MCMC chain}
\usage{
\S4method{length}{SimInf_pmcmc}(x)
}
\arguments{
\item{x}{A \code{SimInf_pmcmc} object containing the MCMC results.}
}
\value{
An integer scalar representing the number of rows in the
    \code{chain} slot (i.e., the total number of samples in the
    chain).
}
\description{
Get the number of iterations (samples) in the Markov Chain Monte
Carlo (MCMC) chain stored in a \code{SimInf_pmcmc} object.
}
