% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_multi_model.R
\name{multi_model}
\alias{multi_model}
\title{Create a multi-model object}
\usage{
multi_model(model, multi_model)
}
\arguments{
\item{model}{The \code{SimInf_model} object to estimate parameters
in.}

\item{multi_model}{FIXME.}
}
\description{
Create a multi-model object
}
