% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.R
\name{n_nodes}
\alias{n_nodes}
\alias{n_nodes,SimInf_model-method}
\alias{n_nodes,SimInf_pfilter-method}
\alias{n_nodes,SimInf_pmcmc-method}
\title{Determine the number of nodes in a model}
\usage{
n_nodes(model)

\S4method{n_nodes}{SimInf_model}(model)

\S4method{n_nodes}{SimInf_pfilter}(model)

\S4method{n_nodes}{SimInf_pmcmc}(model)
}
\arguments{
\item{model}{A \code{SimInf_model} object.}
}
\value{
An integer scalar representing the total number of nodes
    in the model.
}
\description{
Extract the number of nodes from a \code{SimInf_model} object.
A node represents a distinct sub-population in the model, but its
definition is determined by the modeller. For example, a node can
represent a cattle herd, a pen within a herd, or even a single
individual, depending on the research question and the scale of
the study. This count is equivalent to the number of rows in the
initial state vector \code{u0}.
}
\examples{
## Create an 'SIR' model with 100 nodes.
u0 <- data.frame(
  S = rep(99, 100),
  I = rep(1, 100),
  R = rep(0, 100)
)

model <- SIR(
  u0 = u0,
  tspan = 1:10,
  beta = 0.16,
  gamma = 0.077
)

## Get the number of nodes.
n_nodes(model)
}
