% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.R
\name{n_replicates}
\alias{n_replicates}
\alias{n_replicates,SimInf_model-method}
\title{Determine the number of replicates in a model}
\usage{
n_replicates(model)

\S4method{n_replicates}{SimInf_model}(model)
}
\arguments{
\item{model}{A \code{SimInf_model} object.}
}
\value{
An integer scalar representing the number of replicates
    in the model.
}
\description{
Extract the number of replicates from a \code{SimInf_model}
object.  Replicates are independent copies of the model state used
by the \strong{particle filter} (\code{\link{pfilter}}).  This
value is set \strong{internally} by \code{pfilter} and is not
specified when creating a standard model. If the model has not
been processed by a filter, the value will be 1.
}
\examples{
## Create a standard 'SIR' model.
u0 <- data.frame(
  S = rep(99, 100),
  I = rep(1, 100),
  R = rep(0, 100)
)

model <- SIR(
  u0 = u0,
  tspan = 1:10,
  beta = 0.16,
  gamma = 0.077
)

## Get the number of replicates (default is 1 for standard
## models).
n_replicates(model)
}
