% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_raster_model.R
\name{ncol,SimInf_raster_model-method}
\alias{ncol,SimInf_raster_model-method}
\title{Number of columns in the landcover raster}
\usage{
\S4method{ncol}{SimInf_raster_model}(x)
}
\description{
Get the number of columns (\code{ncol}) of the landcover raster.
}
