% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_controller.R
\docType{class}
\name{NULL_controller-class}
\alias{NULL_controller-class}
\title{Definition of the \code{NULL_controller} object for testing}
\description{
Definition of the \code{NULL_controller} object for testing
}
