% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_controller.R
\name{NULL_controller}
\alias{NULL_controller}
\title{Create a \code{NULL_controller} object for testing}
\usage{
NULL_controller(model)
}
\description{
Create a \code{NULL_controller} object for testing
}
