% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/package_skeleton.R
\docType{methods}
\name{package_skeleton}
\alias{package_skeleton}
\alias{package_skeleton,SimInf_mparse-method}
\title{Describe your model in a logical way in R. \code{mparse} creates a
\code{\linkS4class{SimInf_mparse}} object with your model
definition that is ready to be initialised with data and then
\code{\link{run}}.
Create a package skeleton for a model depending on SimInf}
\usage{
package_skeleton(model, name = NULL, path = ".", author = NULL,
  email = NULL, maintainer = NULL, license = "GPL-3")

\S4method{package_skeleton}{SimInf_mparse}(model, name = NULL, path = ".",
  author = NULL, email = NULL, maintainer = NULL, license = "GPL-3")
}
\arguments{
\item{model}{The \code{model} \code{\linkS4class{SimInf_mparse}}
object with your model to create the package skeleton from.}

\item{name}{Character string: the package name and directory name
for your package.}

\item{path}{Path to put the package directory in. Default is '.'
i.e. the current directory.}

\item{author}{Author of the package.}

\item{email}{Email of the package maintainer.}

\item{maintainer}{Maintainer of the package.}

\item{license}{License of the package. Default is 'GPL-3'.}
}
\value{
invisible \code{NULL}.
}
\description{
Describe your model in a logical way in R. \code{mparse} creates a
\code{\linkS4class{SimInf_mparse}} object with your model
definition that is ready to be initialised with data and then
\code{\link{run}}.
Create a package skeleton for a model depending on SimInf
}
\references{
Read the \emph{Writing R Extensions} manual for more
    details.

Once you have created a \emph{source} package you need to install
it: see the \emph{R Installation and Administration} manual,
\code{\link{INSTALL}} and \code{\link{install.packages}}.
}
\keyword{methods}
