% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_skeleton.R
\name{package_skeleton}
\alias{package_skeleton}
\title{Create a package skeleton from a \code{SimInf_model}}
\usage{
package_skeleton(
  model,
  name = NULL,
  path = ".",
  author = NULL,
  email = NULL,
  maintainer = NULL,
  license = "GPL-3"
)
}
\arguments{
\item{model}{A \code{\linkS4class{SimInf_model}} object to create
the package skeleton from.}

\item{name}{Character string with the package name. It must
contain only ASCII letters, numbers, and dots; have at least
two characters; start with a letter; and not end in a dot.
The package name is also used for the class name of the model
and the directory name of the package.}

\item{path}{Path to put the package directory in. Default is
\code{"."} (the current directory).}

\item{author}{Author of the package.}

\item{email}{Email address of the package maintainer.}

\item{maintainer}{Maintainer of the package.}

\item{license}{License of the package. Default is \code{"GPL-3"}.}
}
\value{
\code{invisible(NULL)}.
}
\description{
Generate a source package directory from a
\code{\linkS4class{SimInf_model}} object, allowing the model to be
installed as a standalone add-on R package. This is useful for
distributing custom models or for improving startup performance by
compiling the C code once during package installation rather than
on the first call to \code{run()}.
}
\details{
The typical workflow is:
\enumerate{
  \item Define the model using \code{\link{mparse}} or
    \code{\link{SimInf_model}}.
  \item Call \code{package_skeleton} to generate the package

  \item Install the package using \code{\link{install.packages}}
    or \code{R CMD INSTALL}.
}
}
\references{
Read the \emph{Writing R Extensions} manual for more
    details on building R packages.
}
\seealso{
\code{\link{mparse}} for creating a \code{SimInf_model}
    object from a model specification.  \code{\link{INSTALL}} and
    \code{\link{install.packages}} for installing the generated
    package.
}
