% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_model.R
\docType{methods}
\name{pairs-methods}
\alias{pairs-methods}
\alias{pairs}
\alias{pairs,SimInf_model-method}
\title{Scatterplot Matrices}
\usage{
\S4method{pairs}{SimInf_model}(x, ...)
}
\arguments{
\item{x}{The \code{model} to plot}

\item{...}{Additional arguments affecting the plot produced.}
}
\description{
A matrix of scatterplots with the number of individuals is
produced. The ijth scatterplot contains ‘x[,i]’ plotted against
‘x[,j]’.
}
