% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SISe.R, R/SISe3.R, R/siminf_model.R
\docType{methods}
\name{plot-methods}
\alias{plot}
\alias{plot,SISe-method}
\alias{plot,SISe3-method}
\alias{plot,siminf_model-method}
\alias{plot-methods}
\title{Plot \code{linkS4class{SISe}}}
\usage{
\S4method{plot}{SISe}(x, y, ...)

\S4method{plot}{SISe3}(x, y, ...)

\S4method{plot}{siminf_model}(x, legend, ...)
}
\arguments{
\item{x}{The \code{model} to plot}

\item{y}{Unused argument}

\item{...}{Additional arguments affecting the plot produced.}

\item{legend}{The character vector to appear in the legend.}
}
\description{
Plot \code{linkS4class{SISe}}

Plot \code{linkS4class{SISe3}}

Plot \code{\linkS4class{siminf_model}}
}

