% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_events.R
\docType{methods}
\name{plot,SimInf_events-method}
\alias{plot,SimInf_events-method}
\title{Display the distribution of scheduled events over time}
\usage{
\S4method{plot}{SimInf_events}(x, frame.plot = FALSE, ...)
}
\arguments{
\item{x}{The events data to plot.}

\item{frame.plot}{Draw a frame around each plot. Default is FALSE.}

\item{...}{Additional arguments affecting the plot}
}
\description{
Display the distribution of scheduled events over time
}
