% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfilter.R
\name{plot,SimInf_pfilter-method}
\alias{plot,SimInf_pfilter-method}
\title{Display the sampled trajectory from a particle filter object}
\usage{
\S4method{plot}{SimInf_pfilter}(
  x,
  y,
  level = 1,
  index = NULL,
  range = 0.5,
  type = "s",
  lwd = 2,
  frame.plot = FALSE,
  legend = TRUE,
  ...
)
}
\arguments{
\item{x}{The \code{SimInf_pfilter} object to plot.}

\item{y}{Character vector or formula with the compartments in the
model to include in the plot. Default includes all
compartments in the model. Can also be a formula that
specifies the compartments that define the cases with a
disease or that have a specific characteristic (numerator),
and the compartments that define the entire population of
interest (denominator). The left-hand-side of the formula
defines the cases, and the right-hand-side defines the
population, for example, \code{I~S+I+R} in a \sQuote{SIR}
model (see \sQuote{Examples}). The \code{.}  (dot) is expanded
to all compartments, for example, \code{I~.}  is expanded to
\code{I~S+I+R} in a \sQuote{SIR} model (see
\sQuote{Examples}).}

\item{level}{The level at which the prevalence is calculated at
each time point in \code{tspan}. 1 (population prevalence):
calculates the proportion of the individuals (cases) in the
population. 2 (node prevalence): calculates the proportion of
nodes with at least one case. 3 (within-node prevalence):
calculates the proportion of cases within each node. Default
is \code{1}.}

\item{index}{Indices specifying the nodes to include when plotting
data. Plot one line for each node. Default (\code{index =
NULL}) is to extract data from all nodes and plot the median
count for the specified compartments.}

\item{range}{Show the quantile range of the count in each
compartment. Default is to show the interquartile range
i.e. the middle 50\% of the count in transparent color. The
median value is shown in the same color. Use \code{range =
0.95} to show the middle 95\% of the count. To display
individual lines for each node, specify \code{range = FALSE}.}

\item{type}{The type of plot to draw. The default \code{type =
"s"} draws stair steps. See base plot for other values.}

\item{lwd}{The line width. Default is \code{2}.}

\item{frame.plot}{a logical indicating whether a box should be
drawn around the plot.}

\item{legend}{a logical indicating whether a legend for the
compartments should be added to the plot. A legend is not
drawn for a prevalence plot.}

\item{...}{Other graphical parameters that are passed on to the
plot function.}
}
\description{
Display the sampled trajectory from a particle filter object
}
