% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot,SimInf_pmcmc-method}
\alias{plot,SimInf_pmcmc-method}
\title{Display the PMCMC posterior distribution}
\usage{
\S4method{plot}{SimInf_pmcmc}(x, y, start = 1, end = length(x), thin = 1, ...)
}
\arguments{
\item{x}{The \code{SimInf_pmcmc} object to plot.}

\item{y}{The trace of all variables and logPost are plotted when
\code{y = "trace"} or \code{y = ~trace}, else the posterior
distributions are plotted. Default is to plot the posterier
distributions.}

\item{start}{The start iteration to remove some burn-in
iterations. Default is \code{start = 1}.}

\item{end}{the last iteration to include. Default is \code{NULL}
which set \code{end} to the last iteration in the chain.}

\item{thin}{keep every \code{thin} iteration after the
\code{start} iteration. Default is \code{thin = 1}, i.e., keep
every iteration.}

\item{...}{Additional arguments affecting the plot.}
}
\description{
Display the (approximate) posterior distributions obtained from
     fitting a particle Markov chain Monte Carlo algorithm, or the
     corresponding trace plots.
}
