% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_tidy_events.R
\name{plot,SimInf_tidy_events-method}
\alias{plot,SimInf_tidy_events-method}
\title{Display the distribution of raw events over time}
\usage{
\S4method{plot}{SimInf_tidy_events}(x, frame.plot = FALSE, ...)
}
\arguments{
\item{x}{The raw events data to plot.}

\item{frame.plot}{Draw a frame around each plot. Default is FALSE.}

\item{...}{Additional arguments affecting the plot}
}
\description{
Display the distribution of raw events over time
}
