% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduled_events.R
\name{plot_scheduled_events}
\alias{plot_scheduled_events}
\title{Plot scheduled events}
\usage{
plot_scheduled_events(x, y, events = c("Exit", "Enter", "Internal transfer",
  "External transfer"), frame.plot, ...)
}
\arguments{
\item{x}{the time points of the events.}

\item{y}{the number of events over time.}

\item{events}{the event type to plot.}

\item{frame.plot}{a logical indicating whether a box should be
drawn around the plot.}

\item{...}{additional arguments affecting the plot.}
}
\description{
Plot scheduled events
}
\keyword{internal}

