% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcmc.R
\name{pmcmc}
\alias{pmcmc}
\alias{pmcmc,SimInf_model-method}
\title{Particle Markov chain Monte Carlo (PMCMC) algorithm}
\usage{
pmcmc(
  model,
  obs_process,
  data,
  priors,
  n_particles,
  n_iterations,
  theta = NULL,
  covmat = NULL,
  adaptmix = 0.05,
  adaptive = 100,
  post_proposal = NULL,
  init_model = NULL,
  post_particle = NULL,
  chain = NULL,
  verbose = FALSE
)

\S4method{pmcmc}{SimInf_model}(
  model,
  obs_process,
  data,
  priors,
  n_particles,
  n_iterations,
  theta = NULL,
  covmat = NULL,
  adaptmix = 0.05,
  adaptive = 100,
  post_proposal = NULL,
  init_model = NULL,
  post_particle = NULL,
  chain = NULL,
  verbose = FALSE
)
}
\arguments{
\item{model}{The \code{SimInf_model} object to estimate parameters
for.}

\item{obs_process}{Specification of the stochastic observation
process. The \code{obs_process} can be specified as a
\code{formula} if the model contains only one node and there
is only one data point for each \code{time} in \code{data}.
The left hand side of the formula must match a column name in
the \code{data} data.frame and the right hand side of the
formula is a character specifying the distribution of the
observation process, for example, \code{Iobs ~ poisson(I)}.
The following distributions are supported: \code{x ~
binomial(size, prob)}, \code{x ~ poisson(rate)} and \code{x ~
uniform(min, max)}. The observation process can also be a
function to evaluate the probability density of the
observations given the simulated states. The first argument
passed to the \code{obs_process} function is the result from a
run of the model and it contains one trajectory with simulated
data for a time-point, where the trajectory contains
\code{n_particles} replicates, see
\code{\link{trajectory,SimInf_model-method}}. The second
argument to the \code{obs_process} function is a
\code{data.frame} containing the rows for the specific
time-point that the function is called for. Note that the
function must return the log of the density.}

\item{data}{A \code{data.frame} holding the time series data.}

\item{priors}{The priors for the parameters to fit. Each prior is
specified with a formula notation, for example, \code{beta ~
uniform(0, 1)} specifies that beta is uniformly distributed
between 0 and 1. Use \code{c()} to provide more than one
prior, for example, \code{c(beta ~ uniform(0, 1), gamma ~
normal(10, 1))}. The following distributions are supported:
\code{gamma}, \code{lognormal}, \code{normal} and
\code{uniform}. All parameters in \code{priors} must be only
in either \code{gdata} or \code{ldata}.}

\item{n_particles}{An integer with the number of particles (> 1)
to use at each timestep.}

\item{n_iterations}{An integer specifying the number of iterations
to run the PMCMC.}

\item{theta}{A named numeric vector with initial parameter values.
Default is \code{NULL}, which triggers sampling from the prior
distribution(s).}

\item{covmat}{A named numeric \code{(npars x npars)} covariance
matrix for the proposal distribution. Default is \code{NULL},
which uses \code{diag((theta/10)^2/npars)}.}

\item{adaptmix}{Numeric mixing proportion (0 < value < 1) for
adaptive covariance proposal. Default: 0.05. Larger values
increase the likelihood of using the fixed initial covariance
instead of the adaptive estimate.}

\item{adaptive}{Integer (>= 0) specifying the iteration at which
to start adaptive covariance updates. Default: 100. Set to 0
to disable adaptive updates.}

\item{post_proposal}{An optional function that, if
non-\code{NULL}, is applied on the model after the proposal
has been set for the model, but before running the particle
filter. The function must accept one argument of type
\code{SimInf_model} with the current model of the fitting
process. This function can be useful to update, for example,
\code{ldata} of the model before running a trajectory with
proposed parameters. The function must return the model object
which is then used in the particle filter.}

\item{init_model}{Optional function applied before each particle
filter run. Must accept a \code{SimInf_model} object and
return the modified model. Useful for setting initial states
(\code{u0}, \code{v0}) before running trajectories with
proposed parameters.}

\item{post_particle}{An optional function that, if non-NULL, is
applied after each completed particle. The function must
accept three arguments: 1) an object of \code{SimInf_pmcmc}
with the current state of the fitting process, 2) an object
\code{SimInf_pfilter} with the last particle and one filtered
trajectory attached, and 3) an integer with the iteration in
the fitting process. This function can be useful to, for
example, monitor, save and inspect intermediate results. Note
that the second \code{SimInf_pfilter} argument, is non-NULL
only for the first particle in the chain, and for accepted
particles.}

\item{chain}{Optional \code{data.frame} or object coercible to
one, containing a previous PMCMC chain to continue from. If
provided, \code{theta} and \code{covmat} must be \code{NULL},
and \code{n_iterations} can be 0.}

\item{verbose}{prints diagnostic messages when \code{TRUE}.
Default is \code{FALSE}. When \code{verbose=TRUE}, information
is printed every 100 iterations. For pmcmc, it is possible to
get information every nth information by specifying
\code{verbose=n}, for example, \code{verbose=1} or
\code{verbose=10}.}
}
\value{
A \code{\linkS4class{SimInf_pmcmc}} object containing the
    fitted parameters and diagnostic information for all
    iterations.
}
\description{
Estimates model parameters using the PMCMC algorithm, which
combines a particle filter for likelihood evaluation with a
Metropolis-Hastings MCMC sampler for parameter estimation.
}
\references{
\Andrieu2010

\Roberts2009
}
\seealso{
\code{\link{continue_pmcmc}} for running additional
    iterations.
}
