% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SISe.R, R/SISe3.R
\docType{methods}
\name{prevalence}
\alias{prevalence}
\alias{prevalence,SISe-method}
\alias{prevalence,SISe3-method}
\title{Prevalence}
\usage{
prevalence(model, ...)

\S4method{prevalence}{SISe}(model, by = 1, ...)

\S4method{prevalence}{SISe3}(model, by = 1, ...)
}
\arguments{
\item{model}{The \code{model} to calculated the prevalence from}

\item{...}{Additional arguments affecting the measure}

\item{by}{The number to increment the sequence of time points
starting from 1. Default is 1, which gives the prevalence at every
time point.}
}
\description{
Calculate the proportion infected individuals
}
\keyword{methods}

