% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SimInf_model.R, R/SIR.R,
%   R/SISe.R, R/SISe3.R, R/SISe3_sp.R, R/SISe_sp.R
\docType{methods}
\name{prevalence}
\alias{prevalence}
\alias{prevalence,SimInf_model,character,character-method}
\alias{prevalence,SIR,missing,missing-method}
\alias{prevalence,SISe,missing,missing-method}
\alias{prevalence,SISe3,missing,missing-method}
\alias{prevalence,SISe3_sp,missing,missing-method}
\alias{prevalence,SISe_sp,missing,missing-method}
\title{Prevalence}
\usage{
prevalence(model, numerator, denominator, type = c("pop", "bnp", "wnp"),
  i = NULL)

\S4method{prevalence}{SimInf_model,character,character}(model, numerator,
  denominator, type = c("pop", "bnp", "wnp"), i = NULL)

\S4method{prevalence}{SIR,missing,missing}(model, type, i)

\S4method{prevalence}{SISe,missing,missing}(model, type, i)

\S4method{prevalence}{SISe3,missing,missing}(model, type, i)

\S4method{prevalence}{SISe3_sp,missing,missing}(model, type, i)

\S4method{prevalence}{SISe_sp,missing,missing}(model, type, i)
}
\arguments{
\item{model}{The \code{model} to calculate the prevalence from.}

\item{numerator}{The compartments in the population with a disease
or a condition.}

\item{denominator}{The compartments that define the entire
population of interest.}

\item{type}{The type of prevalence measure to calculate:
\code{'pop'} (default) calcalates the proportion of the
individuals in the population that have disease (model
specific) at each time point in \code{tspan}, \code{'bnp'}
calculates the between-node prevalence, and \code{'wnp'}
calculates the within-node prevalence.}

\item{i}{Indices specifying the nodes to include in the
calculation of the prevalence. Default is \code{NULL}, which
includes all nodes.}
}
\value{
Vector when type equals \code{'pop'} or \code{'bnp'} but
    matrix when type equals \code{'wnp'}.
}
\description{
Calculate the proportion of individuals with disease, or the
proportion of nodes with individuals with disease, or the
proportion of individuals with disease in each node.
}
\section{Methods (by class)}{
\itemize{
\item \code{model = SIR,numerator = missing,denominator = missing}: Prevalence \eqn{= I / (S + I + R)}

\item \code{model = SISe,numerator = missing,denominator = missing}: Prevalence \eqn{= I / (S + I)}

\item \code{model = SISe3,numerator = missing,denominator = missing}: Prevalence \eqn{= (I_1 + I_2 + I_3) / (S_1 + S_2 + S_3 + I_1 + I_2 + I_3)}

\item \code{model = SISe3_sp,numerator = missing,denominator = missing}: Prevalence \eqn{= (I_1 + I_2 + I_3) / (S_1 + S_2 + S_3 + I_1 + I_2 + I_3)}

\item \code{model = SISe_sp,numerator = missing,denominator = missing}: Prevalence \eqn{= I / (S + I)}
}}

