% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/run.r
\name{run_internal}
\alias{run_internal}
\title{Internal function to run siminf stochastic simulation algorithms}
\usage{
run_internal(model, verbose, seed, solver)
}
\arguments{
\item{model}{The siminf model to run.}

\item{verbose}{Level of siminf feeedback during simulation. Silent
if 0, progress if 1, progress and number of transition
events if 2.}

\item{seed}{Random number seed.}

\item{solver}{A character string giving the name of the C function
to initiate and run the model.}
}
\value{
\code{siminf_model} with result from simulation.
}
\description{
Internal function to run siminf stochastic simulation algorithms
}
\keyword{internal}

