% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduled_events.R
\docType{class}
\name{scheduled_events-class}
\alias{scheduled_events-class}
\title{Class \code{"scheduled_events"}}
\description{
Class to handle the scheduled events
}
\section{Slots}{

\describe{
  \item{E}{
    Sparse matrix of object class \code{"\linkS4class{dgCMatrix}"}.
    Each row corresponds to one state in the compartment model. The
    non-zero entries in a column indicates the states to select and
    include in an event.  For the events \code{EXIT_EVENT},
    \code{INTERNAL_TRANSFER_EVENT} and \code{EXTERNAL_TRANSFER_EVENT},
    a non-zero entry in element \code{i} of select column \code{j}
    indicate the compartments to sample individuals from.  For the
    \code{ENTER_EVENT}, all individuals enter first non-zero compartment,
    i.e. a non-zero entry in element \code{i} of select column \code{j}.
  }
  \item{N}{
    For the \code{INTERNAL_TRANSFER_EVENT}, a non-zero entry in
    element \code{i} of select column \code{j} in \code{E} indicate
    the compartments to sample individuals from. The value of element
    \code{i} of column \code{k} in \code{N} determines the
    target compartment. The target compartment for sampled individuals
    is given by adding the value.
  }
  \item{event}{
    Integer vector of length \code{len} with scheduled events.
    The following four type events exists; \code{EXIT_EVENT = 0},
    \code{ENTER_EVENT = 1}, \code{INTERNAL_TRANSFER_EVENT = 2}, and
    \code{EXTERNAL_TRANSFER_EVENT = 3}.
  }
  \item{time}{
    Time for the event. Integer vector of length \code{len}.
  }
  \item{node}{
    The node that the event operates on. Also the source node for
    an \emph{external transfer} event.  Integer vector of length
    \code{len}. 1 <= \code{node[i]} <= Number of nodes.
  }
  \item{dest}{
    The destination node for an \emph{external transfer} event.
    Should be \code{0} for the other event types.
    Integer vector of length \code{len}. dest[i] >= 0.
  }
  \item{n}{
    The number of individuals affected by the event. Integer vector
    of length \code{len}. n[i] >= 0.
  }
  \item{proportion}{
    If \code{n[i]} equals zero, the number of individuals affected by
    \code{event[i]} is calculated by summing the number of individuls
    with \code{proportion[i]}. Numeric vector of length \code{len}.
    0 <= proportion[i] <= 1.
  }
  \item{select}{
    Column \code{j} in the event matrix \code{E} that determines
    the compartments that the event operates on. Integer vector
    of length \code{len}.
  }
  \item{shift}{
    Column \code{k} in the shift matrix \code{N} that determines
    how individuals in \emph{internal transfer} and \emph{external
    transfer} events are shifted to enter another compartment.
    Should be \code{0} for the other event types.
    Integer vector of length \code{len}.
  }
}
}
\keyword{methods}

