% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR.R
\docType{class}
\name{SEIR-class}
\alias{SEIR-class}
\title{Definition of the \acronym{SEIR} model}
\description{
Class to handle the \acronym{SEIR} \code{\link{SimInf_model}}.
}
\seealso{
\code{\link{SEIR}} for creating an \acronym{SEIR} model
}
