% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/external_events.R
\docType{methods}
\name{show,external_events-method}
\alias{show,external_events-method}
\alias{show,external_events-methods}
\title{Brief summary of \code{external_events}}
\usage{
\S4method{show}{external_events}(object)
}
\arguments{
\item{object}{The external_events \code{object}}
}
\value{
None (invisible 'NULL').
}
\description{
Brief summary of \code{external_events}
}
\keyword{methods}

