% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduled_events.R
\docType{methods}
\name{show,scheduled_events-method}
\alias{show,scheduled_events-method}
\alias{show,scheduled_events-methods}
\title{Brief summary of \code{scheduled_events}}
\usage{
\S4method{show}{scheduled_events}(object)
}
\arguments{
\item{object}{The scheduled_events \code{object}}
}
\value{
None (invisible 'NULL').
}
\description{
Shows the number of scheduled events.
}
\keyword{methods}
