% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_smc.R
\name{show,SimInf_abc_smc-method}
\alias{show,SimInf_abc_smc-method}
\title{Print summary of a \code{SimInf_abc_smc} object}
\usage{
\S4method{show}{SimInf_abc_smc}(object)
}
\arguments{
\item{object}{The \code{SimInf_abc_smc} object.}
}
\value{
\code{invisible(object)}.
}
\description{
Print summary of a \code{SimInf_abc_smc} object
}
