% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterated-filtering.R
\name{show,SimInf_if2-method}
\alias{show,SimInf_if2-method}
\title{Brief summary of a \code{SimInf_if2} object}
\usage{
\S4method{show}{SimInf_if2}(object)
}
\arguments{
\item{object}{The \code{SimInf_if2} object.}
}
\value{
\code{invisible(object)}.
}
\description{
Brief summary of a \code{SimInf_if2} object
}
