% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_individual_events.R
\name{show,SimInf_indiv_events-method}
\alias{show,SimInf_indiv_events-method}
\title{Print summary of a \code{SimInf_indiv_events} object}
\usage{
\S4method{show}{SimInf_indiv_events}(object)
}
\arguments{
\item{object}{The \code{SimInf_indiv_events} object.}
}
\value{
\code{invisible(object)}.
}
\description{
Print summary of a \code{SimInf_indiv_events} object
}
