% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfilter.R
\name{show,SimInf_pfilter-method}
\alias{show,SimInf_pfilter-method}
\title{Brief summary of a \code{SimInf_pfilter} object}
\usage{
\S4method{show}{SimInf_pfilter}(object)
}
\arguments{
\item{object}{The \code{\linkS4class{SimInf_pfilter}} object to
display.}
}
\value{
The \code{object}, returned invisibly.
}
\description{
Display a brief summary of a bootstrap particle filter analysis,
including the number of particles used and the estimated
log-likelihood.
}
\seealso{
\code{\link{pfilter}} for running a particle filter,
    \code{\link{logLik}} for extracting the log-likelihood,
    \code{\linkS4class{SimInf_pfilter}} for the class definition.
}
