% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcmc.R
\name{show,SimInf_pmcmc-method}
\alias{show,SimInf_pmcmc-method}
\title{Brief summary of a \code{SimInf_pmcmc} object}
\usage{
\S4method{show}{SimInf_pmcmc}(object)
}
\arguments{
\item{object}{The \code{\linkS4class{SimInf_pmcmc}} object to
display.}
}
\value{
The \code{object}, returned invisibly.
}
\description{
Display a summary of a particle Markov chain Monte Carlo (PMCMC)
analysis, including the number of iterations, number of particles,
mixing proportion for the adaptive proposal, acceptance ratio, and
posterior quantiles (2.5%, 25%, 50%, 75%, 97.5%), mean, and
standard deviation for each estimated parameter.
}
\seealso{
\code{\link{pmcmc}} for running a PMCMC analysis,
    \code{\link{continue_pmcmc}} for continuing an existing PMCMC
    chain, \code{\linkS4class{SimInf_pmcmc}} for the class
    definition.
}
