% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_events.R
\name{show,SimInf_raw_events-method}
\alias{show,SimInf_raw_events-method}
\title{Print summary of a \code{SimInf_raw_events} object}
\usage{
\S4method{show}{SimInf_raw_events}(object)
}
\arguments{
\item{object}{The \code{SimInf_raw_events} object.}
}
\value{
\code{invisible(object)}.
}
\description{
Print summary of a \code{SimInf_raw_events} object
}
