% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_smc.R
\docType{class}
\name{SimInf_abc_smc-class}
\alias{SimInf_abc_smc-class}
\title{Class \code{"SimInf_abc_smc"}}
\description{
Class \code{"SimInf_abc_smc"}
}
\section{Slots}{

\describe{
  \item{model}{
    FIXME.
  }
  \item{priors}{
    FIXME.
  }
  \item{i}{
    FIXME.
  }
  \item{fn}{
    FIXME.
  }
  \item{x}{
    FIXME.
  }
  \item{w}{
    FIXME.
  }
}
}

