% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_controller.R
\name{SimInf_controller}
\alias{SimInf_controller}
\title{Create a \code{SimInf_controller}}
\usage{
SimInf_controller(model = NULL, C_code = NULL)
}
\arguments{
\item{model}{The \code{linkS4class{SimInf_model}} object to apply
the controller on.}

\item{C_code}{Character vector with optional controller C code. If
non-empty, the C code is written to a temporary C-file when
the \code{run} method is called.  The temporary C-file is
compiled and the resulting DLL is dynamically loaded. The DLL
is unloaded and the temporary files are removed after running
the controller.}
}
\value{
\linkS4class{SimInf_controller} object.
}
\description{
Create a \code{SimInf_controller}
}
