% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_indiv_events.R
\name{[,SimInf_indiv_events,ANY,ANY,ANY-method}
\alias{[,SimInf_indiv_events,ANY,ANY,ANY-method}
\title{Extract individuals from tidy events}
\usage{
\S4method{[}{SimInf_indiv_events,ANY,ANY,ANY}(x, i, j)
}
\arguments{
\item{x}{a tidy events \code{object}.}

\item{i}{FIXME.}

\item{j}{FIXME.}
}
\value{
a \code{data.frame} with the columns \code{id},
    \code{node}, and \code{age}.
}
\description{
Lookup individuals, in which node they are located and their age
at a specified time-point \code{i}.
}
