% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_model.R
\docType{class}
\name{SimInf_model_dll-class}
\alias{SimInf_model_dll-class}
\title{Class to handle a compiled custom \code{\link{SimInf_model}}}
\description{
Class to handle a compiled custom \code{\link{SimInf_model}}
}
\section{Slots}{

\describe{
  \item{filename}{
    Character vector of length 1 containing name of shared library.
  }
}
}

