% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_multi_model.R
\docType{class}
\name{SimInf_multi_model-class}
\alias{SimInf_multi_model-class}
\title{Class \code{"SimInf_multi_model"}}
\description{
Class \code{"SimInf_multi_model"}
}
\section{Slots}{

\describe{
\item{\code{model}}{The \code{SimInf_model} object to estimate parameters
in.}

\item{\code{multi_model}}{FIXME.}

\item{\code{n_models}}{Integer with the number of models.}
}}

