% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcmc.R
\docType{class}
\name{SimInf_pmcmc-class}
\alias{SimInf_pmcmc-class}
\title{Class \code{SimInf_pmcmc}}
\description{
Class \code{SimInf_pmcmc}
}
\section{Slots}{

\describe{
\item{\code{model}}{The \code{SimInf_model} object to estimate parameters
in.}

\item{\code{priors}}{A \code{data.frame} defining the prior distributions
for the parameters. It contains four columns:
\itemize{
  \item \code{parameter}: The name of the parameter in the
  model.
  \item \code{distribution}: The prior distribution
    type. Valid values are \code{"gamma"}, \code{"lognormal"},
    \code{"normal"}, or \code{"uniform"}.
  \item \code{p1}: The first hyperparameter:
    \itemize{
      \item \code{"gamma"}: \emph{shape}
      \item \code{"lognormal"}: \emph{meanlog} (mean on the
      log scale)
      \item \code{"normal"}: \emph{mean}
      \item \code{"uniform"}: \emph{lower} bound
    }
  \item \code{p2}: The second hyperparameter:
    \itemize{
      \item \code{"gamma"}: \emph{rate}
      \item \code{"lognormal"}: \emph{sdlog} (standard
      deviation on the log scale)
      \item \code{"normal"}: \emph{sd} (standard deviation)
      \item \code{"uniform"}: \emph{upper} bound
    }
}}

\item{\code{target}}{Character vector (\code{"gdata"} or \code{"ldata"})
that determines if the ABC-SMC method estimates parameters in
\code{model@gdata} (global data) or in \code{model@ldata}
(local data).}

\item{\code{pars}}{An integer vector with the indices of the parameters in
\code{target} that are being estimated.}

\item{\code{n_particles}}{An integer with the number of particles (> 1) to
use in the bootstrap particle filter.}

\item{\code{data}}{A \code{data.frame} holding the time series data for
the observation process.}

\item{\code{chain}}{A matrix where each row contains \code{logPost},
\code{logLik}, \code{logPrior}, \code{accept}, and the
\code{parameters} for each iteration.}

\item{\code{covmat}}{A named numeric \code{(npars x npars)} matrix with
covariances to use as initial proposal matrix.}

\item{\code{adaptmix}}{A numeric scalar specifying the mixing proportion
for the adaptive proposal distribution.}

\item{\code{adaptive}}{An integer specifying when to start the adaptive
update of the proposal distribution (iteration number).}
}}

\seealso{
\code{\link{pmcmc}} for the main PMCMC function,
    \code{\link{continue_pmcmc}} for continuing an existing PMCMC
    run.
}
