% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{SimInf_raster_model-class}
\alias{SimInf_raster_model-class}
\title{Class \code{"SimInf_raster_model"}}
\description{
Class to handle data for a \code{SimInf_raster_model} object.  It
is a model where the nodes are not fixed at one position but can
move between cells on a raster.
}
\section{Slots}{

\describe{
\item{\code{cell_S}}{FIXME.}

\item{\code{raster}}{An integer matrix that represents a spatially
referenced surface divided into two dimensional cells (rows,
and columns). In the raster, cell numbers start at 1 in the
upper left corner, and increase from left to right, and then
from top to bottom. The last cell number equals the number of
cells of the raster. Row numbers start at 1 at the top, column
numbers start at 1 at the left.}
}}

