% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_raw_events.R
\docType{class}
\name{SimInf_raw_events-class}
\alias{SimInf_raw_events-class}
\title{Class \code{"SimInf_raw_events"}}
\description{
Class \code{"SimInf_raw_events"}
}
\section{Slots}{

\describe{
\item{\code{id}}{FIXME}

\item{\code{event}}{FIXME}

\item{\code{time}}{FIXME}

\item{\code{node}}{FIXME}

\item{\code{dest}}{FIXME}

\item{\code{keep}}{FIXME}
}}

