% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_tidy_events.R
\name{[,SimInf_tidy_events,integer,missing,ANY-method}
\alias{[,SimInf_tidy_events,integer,missing,ANY-method}
\alias{[,SimInf_tidy_events,numeric,missing,ANY-method}
\title{Extract individuals from tidy events}
\usage{
\S4method{[}{SimInf_tidy_events,integer,missing,ANY}(x, i)

\S4method{[}{SimInf_tidy_events,numeric,missing,ANY}(x, i)
}
\arguments{
\item{x}{a tidy events \code{object}.}

\item{i}{FIXME.}
}
\value{
a \code{data.frame} with the columns \code{id},
    \code{node}, and \code{age}.
}
\description{
Lookup individuals, in which node they are located and their age
at a specified time-point \code{i}.
}
