% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SISe.R
\docType{class}
\name{SISe-class}
\alias{SISe-class}
\title{Class SISe}
\description{
Class to handle the \acronym{SISe} model. This class inherits from
\code{\linkS4class{SimInf_model}}, meaning that \acronym{SIR}
objects are fully compatible with all generic functions defined
for \code{SimInf_model}, such as \code{\link{run}},
\code{\link{plot}}, \code{\link{trajectory}}, and
\code{\link{prevalence}}.
}
\seealso{
\code{\link{SISe}} for creating an SISe model
}
