% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SISe3_sp.R
\docType{class}
\name{SISe3_sp-class}
\alias{SISe3_sp-class}
\title{Class SISe3_sp}
\description{
Class to handle the \acronym{SISe3_sp} model. This class inherits
from \code{\linkS4class{SimInf_model}}, meaning that
\acronym{SISe3_sp} objects are fully compatible with all generic
functions defined for \code{SimInf_model}, such as
\code{\link{run}}, \code{\link{plot}}, \code{\link{trajectory}},
and \code{\link{prevalence}}.
}
