% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduled_events.R
\docType{methods}
\name{summary,scheduled_events-method}
\alias{summary,scheduled_events-method}
\alias{summary,scheduled_events-methods}
\title{Summary of \code{scheduled_events}}
\usage{
\S4method{summary}{scheduled_events}(object, ...)
}
\arguments{
\item{object}{The \code{scheduled_events} object}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
None (invisible 'NULL').
}
\description{
Shows the number of scheduled events and the number of scheduled
events per event type.
}
\keyword{methods}

