% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_smc.R
\name{summary,SimInf_abc_smc-method}
\alias{summary,SimInf_abc_smc-method}
\title{Detailed summary of a \code{SimInf_abc_smc} object}
\usage{
\S4method{summary}{SimInf_abc_smc}(object, ...)
}
\arguments{
\item{object}{The \code{SimInf_abc_smc} object}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
None (invisible 'NULL').
}
\description{
Detailed summary of a \code{SimInf_abc_smc} object
}
