% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterated-filtering.R
\name{summary,SimInf_if2-method}
\alias{summary,SimInf_if2-method}
\title{Detailed summary of a \code{SimInf_if2} object}
\usage{
\S4method{summary}{SimInf_if2}(object, ...)
}
\arguments{
\item{object}{The \code{SimInf_if2} object}

\item{...}{Not used.}
}
\value{
None (invisible 'NULL').
}
\description{
Detailed summary of a \code{SimInf_if2} object
}
