% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/measures.r
\docType{methods}
\name{susceptible}
\alias{susceptible}
\alias{susceptible,SISe3-method}
\title{Susceptible}
\usage{
susceptible(model, ...)

\S4method{susceptible}{SISe3}(model, age = c("age_1", "age_2", "age_3"), ...)
}
\arguments{
\item{model}{The \code{model} to extract the susceptible from}

\item{...}{Additional arguments affecting the measure}

\item{age}{The age category to extract
##'}
}
\description{
Extracts the number of susceptible
}
\keyword{methods}

