% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_individual_events.R
\name{as.data.frame.SimInf_individual_events}
\alias{as.data.frame.SimInf_individual_events}
\title{Coerce to data frame}
\usage{
\method{as.data.frame}{SimInf_individual_events}(x, ...)
}
\arguments{
\item{x}{any \R object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
Coerce to data frame
}
