% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_smc.R
\name{continue}
\alias{continue}
\title{Run more generations of ABC SMC}
\usage{
continue(object, ngen = 1, ..., verbose = TRUE)
}
\arguments{
\item{object}{The \code{SimInf_abc_smc} to continue from.}

\item{ngen}{The number of generations of ABC-SMC to run.}

\item{...}{Further arguments to be passed to
\code{SimInf_abc_smc@fn}.}

\item{verbose}{= TRUE}
}
\value{
A \code{SimInf_abc_smc} object.
}
\description{
Run more generations of ABC SMC
}
