% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R, R/pmcmc.R
\name{continue}
\alias{continue}
\alias{continue,SimInf_abc-method}
\alias{continue,SimInf_pmcmc-method}
\title{Run more generations of ABC SMC}
\usage{
continue(object, ..., verbose = getOption("verbose", FALSE))

\S4method{continue}{SimInf_abc}(object, ngen = 1, ..., verbose = getOption("verbose", FALSE))

\S4method{continue}{SimInf_pmcmc}(object, niter, ..., verbose = getOption("verbose", FALSE))
}
\arguments{
\item{object}{The \code{SimInf_abc} to continue from.}

\item{...}{Further arguments to be passed to
\code{SimInf_abc@fn}.}

\item{verbose}{prints diagnostic messages when \code{TRUE}. The
default is to retrieve the global option \code{verbose} and
use \code{FALSE} if it is not set.}

\item{ngen}{The number of generations of ABC-SMC to run.}
}
\value{
A \code{SimInf_abc} object.
}
\description{
Run more generations of ABC SMC
}
