% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{continue}
\alias{continue}
\title{Run more generations of ABC SMC}
\usage{
continue(object, ngen = 1, ..., verbose = getOption("verbose", FALSE))
}
\arguments{
\item{object}{The \code{SimInf_abc} to continue from.}

\item{ngen}{The number of generations of ABC-SMC to run.}

\item{...}{Further arguments to be passed to
\code{SimInf_abc@fn}.}

\item{verbose}{prints diagnostic messages when \code{TRUE}. The
default is to retrieve the global option \code{verbose} and
use \code{FALSE} if it is not set.}
}
\value{
A \code{SimInf_abc} object.
}
\description{
Run more generations of ABC SMC
}
