% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcmc.R
\name{continue_pmcmc}
\alias{continue_pmcmc}
\alias{continue_pmcmc,SimInf_pmcmc-method}
\title{Run more iterations of PMCMC}
\usage{
continue_pmcmc(
  object,
  n_iterations,
  init_model = NULL,
  post_particle = NULL,
  verbose = getOption("verbose", FALSE)
)

\S4method{continue_pmcmc}{SimInf_pmcmc}(
  object,
  n_iterations,
  init_model = NULL,
  post_particle = NULL,
  verbose = getOption("verbose", FALSE)
)
}
\arguments{
\item{object}{The \code{SimInf_pmcmc} object to continue from.}

\item{n_iterations}{An integer specifying the number of iterations
to run the PMCMC.}

\item{init_model}{An optional function that, if non-NULL, is
applied before running each proposal. The function must accept
one argument of type \code{SimInf_model} with the current
model of the fitting process. This function can be useful to
specify the initial state of \code{u0} or \code{v0} of the
model before running a trajectory with proposed parameters.}

\item{post_particle}{An optional function that, if non-NULL, is
applied after each completed particle. The function must
accept three arguments: 1) an object of \code{SimInf_pmcmc}
with the current state of the fitting process, 2) an object
\code{SimInf_pfilter} with the last particle and one filtered
trajectory attached, and 3) an integer with the iteration in
the fitting process. This function can be useful to, for
example, monitor, save and inspect intermediate results. Note
that the second \code{SimInf_pfilter} argument, is non-NULL
only for the first particle in the chain, and for accepted
particles.}

\item{verbose}{prints diagnostic messages when \code{TRUE}. The
default is to retrieve the global option \code{verbose} and
use \code{FALSE} if it is not set. When \code{verbose=TRUE},
information is printed every 100 iterations. For pmcmc, it is
possible to get information every nth information by
specifying \code{verbose=n}, for example, \code{verbose=1} or
\code{verbose=10}.}
}
\description{
Run more iterations of PMCMC
}
