% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcmc.R
\name{continue_pmcmc}
\alias{continue_pmcmc}
\alias{continue_pmcmc,SimInf_pmcmc-method}
\title{Run more iterations of PMCMC}
\usage{
continue_pmcmc(object, ...)

\S4method{continue_pmcmc}{SimInf_pmcmc}(object, niter, ..., verbose = getOption("verbose", FALSE))
}
\arguments{
\item{object}{The \code{SimInf_pmcmc} object to continue from.}

\item{...}{Unused additional arguments.}

\item{niter}{An integer specifying the number of iterations to run
the PMCMC.}

\item{verbose}{prints diagnostic messages when \code{TRUE}. The
default is to retrieve the global option \code{verbose} and
use \code{FALSE} if it is not set. When \code{verbose=TRUE},
information is printed every 100 iterations. For pmcmc, it is
possible to get information every nth information by
specifying \code{verbose=n}, for example, \code{verbose=1} or
\code{verbose=10}.}
}
\description{
Run more iterations of PMCMC
}
