% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/demo_model.R
\name{demo_model}
\alias{demo_model}
\title{Generate a model for demonstration}
\usage{
demo_model(nodes = 1, days = 1000, model = c("SISe", "SISe3"))
}
\arguments{
\item{nodes}{Number of nodes in the model. Default is 1.}

\item{days}{Number of days to model. Default is 1000.}

\item{model}{The name of the model. Default is 'SISe'.}
}
\value{
A model
}
\description{
Generate a model for demonstration
}

