% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_raster_model.R
\name{dim,SimInf_raster_model-method}
\alias{dim,SimInf_raster_model-method}
\title{Dimension of the landcover raster.}
\usage{
\S4method{dim}{SimInf_raster_model}(x)
}
\arguments{
\item{x}{\code{SimInf_raster_model} object.}
}
\value{
integer vector with the number of rows and columns.
}
\description{
Get the dimension (\code{dim}) of the landcover raster.
}
