% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance_matrix}
\alias{distance_matrix}
\title{Create distance matrix}
\usage{
distance_matrix(x, y, cutoff)
}
\arguments{
\item{x}{Projected x coordinate}

\item{y}{Projected y coordinate}

\item{cutoff}{The distance cutoff}
}
\value{
\code{dgCMatrix}
}
\description{
Calculate the euclidian distances beween coordinates for
coordinates within the cutoff.
}
\examples{
distance_matrix(1:10, 1:10, 3)
}

