% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance_matrix}
\alias{distance_matrix}
\title{Create distance matrix}
\usage{
distance_matrix(x, y, cutoff, min_dist = NULL)
}
\arguments{
\item{x}{Projected x coordinate}

\item{y}{Projected y coordinate}

\item{cutoff}{The distance cutoff}

\item{min_dist}{The minimum distance to separate two nodes.  If
the coordinates for two nodes are identical, the min_dist must
be assigned or an error is raised.  This is to protect against
infinite values when calculating \eqn{1 / d^2}. Default is
NULL.}
}
\value{
\code{dgCMatrix}
}
\description{
Calculate the euclidian distances beween coordinates for all
coordinates within the cutoff.
}
\examples{
distance_matrix(1:10, 1:10, 3)
}
