% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_events.R, R/SimInf_indiv_events.R
\name{events}
\alias{events}
\alias{events,SimInf_model-method}
\alias{events,SimInf_indiv_events-method}
\title{Extract the events from a \code{SimInf_model} object}
\usage{
events(object, ...)

\S4method{events}{SimInf_model}(object, ...)

\S4method{events}{SimInf_indiv_events}(object, time = NULL, target = NULL, age = NULL)
}
\arguments{
\item{object}{The \code{model} to extract the events from.}

\item{...}{Additional arguments affecting the generated events.}

\item{time}{Only used when object is of class
\code{SimInf_indiv_events} object. All events that occur after
\sQuote{time} are included. Default is \code{NULL} which means
to extract the events after the minimum time-point in the
\code{SimInf_indiv_events} object.}

\item{target}{the SimInf model ('SEIR', 'SIR', 'SIS', 'SISe3',
'SISe3_sp', 'SISe', or 'SISe_sp') to target the events and u0
for. The default, \code{NULL}, creates events but they might
have to be post-processed to fit the specific use case.}

\item{age}{FIXME.}
}
\value{
\code{\linkS4class{SimInf_events}} object.
}
\description{
Extract the scheduled events from a \code{SimInf_model} object.
}
\examples{
## Create an SIR model that includes scheduled events.
model <- SIR(u0     = u0_SIR(),
             tspan  = 1:(4 * 365),
             events = events_SIR(),
             beta   = 0.16,
             gamma  = 0.077)

## Extract the scheduled events from the model and display summary
summary(events(model))

## Extract the scheduled events from the model and plot them
plot(events(model))
}
