% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR.R
\name{events_SEIR}
\alias{events_SEIR}
\title{Example data with scheduled events for the \code{SEIR} model}
\usage{
events_SEIR()
}
\value{
A \code{data.frame}
}
\description{
Synthetic scheduled events data to demonstrate the \code{SEIR}
model. The data contains 466692 events for 1600 nodes over 365 * 4
days.
}
