% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR.R
\name{events_SIR}
\alias{events_SIR}
\title{Example data with scheduled events for the \code{SIR} model}
\usage{
events_SIR()
}
\value{
A \code{data.frame}
}
\description{
Synthetic scheduled events data to demonstrate the \code{SIR}
model. The data contains 466692 events for 1600 nodes over 365 * 4
days.
}
