% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SISe.R
\name{events_SISe}
\alias{events_SISe}
\title{Scheduled events example data for the \code{SISe} model}
\usage{
events_SISe()
}
\value{
A \code{data.frame}
}
\description{
Synthetic scheduled events data to demonstrate the \code{SISe}
model. The data contains 466692 events for 1600 nodes over 365 * 4
days.
}
\keyword{methods}
