% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf.R
\docType{data}
\name{events_SISe3}
\alias{events_SISe3}
\title{Scheduled events example data}
\format{A \code{data.frame}}
\usage{
data(events_SISe3)
}
\description{
Synthetic scheduled events data to demonstrate the \code{SISe3}
model. The data contains 783773 events for 1600 nodes over 365 * 4
days.
}
\keyword{dataset}
