% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterated-filtering.R
\name{if2}
\alias{if2}
\alias{if2,SimInf_model-method}
\title{Iterated filtering algorithm}
\usage{
if2(model, obs_process, n_iterations, verbose = getOption("verbose", FALSE))

\S4method{if2}{SimInf_model}(model, obs_process, n_iterations, verbose = getOption("verbose", FALSE))
}
\arguments{
\item{model}{The model to simulate data from.}

\item{obs_process}{Specification of the stochastic observation
process. The \code{obs_process} can be specified as a
\code{formula} if the model contains only one node and there
is only one data point for each \code{time} in \code{data}.
The left hand side of the formula must match a column name in
the \code{data} data.frame and the right hand side of the
formula is a character specifying the distribution of the
observation process, for example, \code{Iobs ~ poisson(I)}.
The following distributions are supported: \code{x ~
binomial(size, prob)}, \code{x ~ poisson(rate)} and \code{x ~
uniform(min, max)}. The observation process can also be a
function to evaluate the probability density of the
observations given the simulated states. The first argument
passed to the \code{obs_process} function is the result from a
run of the model and it contains one trajectory with simulated
data for a time-point, where the trajectory contains
\code{n_particles} replicates, see
\code{\link{trajectory,SimInf_model-method}}. The second
argument to the \code{obs_process} function is a
\code{data.frame} containing the rows for the specific
time-point that the function is called for. Note that the
function must return the log of the density.}

\item{n_iterations}{An integer specifying the number of filtering
iterations to perform.}

\item{verbose}{prints diagnostic messages when \code{TRUE}. The
default is to retrieve the global option \code{verbose} and
use \code{FALSE} if it is not set.}
}
\description{
Iterated filtering algorithm
}
\references{
\Ionides2015
}
