% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SISe.R, R/SISe3.R
\docType{methods}
\name{infected}
\alias{infected}
\alias{infected,SISe-method}
\alias{infected,SISe3-method}
\title{Infected}
\usage{
infected(model, ...)

\S4method{infected}{SISe}(model, ...)

\S4method{infected}{SISe3}(model, age = c("age_1", "age_2", "age_3"), ...)
}
\arguments{
\item{model}{The \code{model} to extract the infected from}

\item{...}{Additional arguments affecting the measure}

\item{age}{The age category to extract}
}
\description{
Extracts the number of infected
}
\keyword{methods}

