% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/external_events.r
\name{is_wholenumber}
\alias{is_wholenumber}
\title{Check if wholenumbers}
\usage{
is_wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{Value to check}

\item{tol}{Tolerance of the check}
}
\value{
logical vector
}
\description{
Check that all values are wholenumbers, see example in integer {base}
}
\keyword{internal}

