% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcmc.R
\name{length,SimInf_pmcmc-method}
\alias{length,SimInf_pmcmc-method}
\title{Length of the MCMC chain}
\usage{
\S4method{length}{SimInf_pmcmc}(x)
}
\arguments{
\item{x}{The \code{SimInf_pmcmc} object determine the length of
the MCMC chain for.}
}
\description{
Length of the MCMC chain
}
