% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfilter.R
\name{logLik,SimInf_pfilter-method}
\alias{logLik,SimInf_pfilter-method}
\title{Log likelihood}
\usage{
\S4method{logLik}{SimInf_pfilter}(object)
}
\arguments{
\item{object}{The \code{SimInf_pfilter} object.}
}
\value{
the estimated log likelihood.
}
\description{
Extract the estimated log likelihood from a \code{SimInf_pfilter}
object.
}
