% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{mparse}
\alias{mparse}
\title{Model parser}
\usage{
mparse(transitions = NULL, compartments = NULL, ...)
}
\arguments{
\item{compartments}{contains the names of the involved
compartments, for example, \code{compartments = c("S", "I",
"R")}.}

\item{...}{rate-constants for the model.}
}
\value{
\linkS4class{SimInf_mparse}
##' @export
}
\description{
Model parser
}
