% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{n_generations}
\alias{n_generations}
\alias{n_generations,SimInf_abc-method}
\title{Determine the number of generations}
\usage{
n_generations(object)

\S4method{n_generations}{SimInf_abc}(object)
}
\arguments{
\item{object}{the \code{SimInf_abc} object to determine the number
of generations for.}
}
\value{
an integer with the number of generations.
}
\description{
Determine the number of generations
}
