% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf.R
\docType{data}
\name{nodes}
\alias{nodes}
\title{Example data with spatial distribution of nodes}
\format{A \code{data.frame}}
\usage{
data(nodes)
}
\description{
Synthetic data with spatial distribution of 1600 nodes to
demonstrate various models.
}
\examples{
\dontrun{
data(nodes)
plot(y ~ x, nodes)
}
}
\keyword{dataset}
