% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_raster_model.R
\name{nrow,SimInf_raster_model-method}
\alias{nrow,SimInf_raster_model-method}
\title{Number of rows in the landcover raster}
\usage{
\S4method{nrow}{SimInf_raster_model}(x)
}
\description{
Get the number of rows (\code{nrow}) of the landcover raster.
}
