% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfilter.R
\name{pfilter}
\alias{pfilter}
\alias{pfilter,SimInf_model-method}
\title{Bootstrap particle filter}
\usage{
pfilter(model, obs_process, data, npart)

\S4method{pfilter}{SimInf_model}(model, obs_process, data, npart)
}
\arguments{
\item{model}{The \code{SimInf_model} object to simulate data from.}

\item{obs_process}{A \code{formula} or \code{function} determining
the observation process.}

\item{data}{A \code{data.frame} holding the time series data.}

\item{npart}{An integer with the number of particles (> 1) to use
at each timestep.}
}
\value{
A \code{SimInf_pfilter} object.
}
\description{
Systematic resampling is performed at each observation.
}
