% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SISe.R
\docType{methods}
\name{plot-methods}
\alias{plot}
\alias{plot,SISe-method}
\alias{plot-methods}
\title{plot,-method}
\usage{
\S4method{plot}{SISe}(x, y, ...)
}
\arguments{
\item{x}{The \code{model} to plot}

\item{y}{Unused argument}

\item{...}{Additional arguments affecting the plot produced.}
}
\description{
plot,-method
}

