% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SISe.R, R/SISe3.R, R/siminf_model.R
\docType{methods}
\name{plot-methods}
\alias{plot}
\alias{plot,SISe-method}
\alias{plot,SISe3-method}
\alias{plot,siminf_model-method}
\alias{plot-methods}
\title{Plot \code{\linkS4class{siminf_model}}}
\usage{
\S4method{plot}{SISe}(x, t0 = NULL, ...)

\S4method{plot}{SISe3}(x, t0 = NULL, ...)

\S4method{plot}{siminf_model}(x, legend, t0 = NULL, ...)
}
\arguments{
\item{x}{The \code{model} to plot}

\item{t0}{The first date of \code{x@tspan} as a character string
in format 'yyyy-mm-dd'. Default is NULL which prints the x-axis
labels as the sequence 1:length(x@tspan). If non-null, the labels
are converted to dates.}

\item{...}{Additional arguments affecting the plot produced.}

\item{legend}{The character vector to appear in the legend.}

\item{y}{Unused argument}
}
\description{
Plot \code{\linkS4class{siminf_model}}
}

