% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{plot,SimInf_abc-method}
\alias{plot,SimInf_abc-method}
\title{Display the ABC posterior distribution}
\usage{
\S4method{plot}{SimInf_abc}(x, y, ...)
}
\arguments{
\item{x}{The \code{SimInf_abc} object to plot.}

\item{y}{The generation to plot. The default is to display the
last generation.}

\item{...}{Additional arguments affecting the plot.}
}
\description{
Display the ABC posterior distribution
}
