% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_individual_events.R
\name{plot,SimInf_indiv_events-method}
\alias{plot,SimInf_indiv_events-method}
\title{Display the distribution of individual events over time}
\usage{
\S4method{plot}{SimInf_indiv_events}(x, frame.plot = FALSE, ...)
}
\arguments{
\item{x}{The individual events data to plot.}

\item{frame.plot}{a logical indicating whether a box should be
drawn around the plot.}

\item{...}{Other graphical parameters that are passed on to the
plot function.}
}
\description{
Display the distribution of individual events over time
}
