% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfilter.R
\name{plot,SimInf_pfilter-method}
\alias{plot,SimInf_pfilter-method}
\title{Diagnostic plot of a particle filter object}
\usage{
\S4method{plot}{SimInf_pfilter}(x, y, ...)
}
\arguments{
\item{x}{The \code{SimInf_pfilter} object to plot.}

\item{y}{If y is \code{NULL} or missing (default), the filtered
trajectory (top) and the effective sample size (bottom) are
displayed. If \code{y} is a character vector or a formula, the
plot function for a \code{SimInf_model} object is called with
the filtered trajectory, see
\code{\link{plot,SimInf_model-method}} for more details about
the specification a plot.}

\item{...}{Other graphical parameters that are passed on to the
plot function.}
}
\description{
Diagnostic plot of a particle filter object
}
