% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfilter.R
\name{plot,SimInf_pfilter-method}
\alias{plot,SimInf_pfilter-method}
\title{Diagnostic plot of a particle filter object}
\usage{
\S4method{plot}{SimInf_pfilter}(x)
}
\arguments{
\item{x}{The \code{SimInf_pfilter} object to plot.}
}
\description{
Displays the randomly selected trajectory from the particle filter
(top), and the effective sample size (bottom).
}
