% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcmc.R
\name{pmcmc}
\alias{pmcmc}
\title{Particle Markov chain Monte Carlo (PMCMC) algorithm}
\usage{
pmcmc(
  model,
  data,
  priors,
  npart,
  niter,
  fn,
  ...,
  verbose = getOption("verbose", FALSE)
)
}
\arguments{
\item{model}{The model to simulate data from.}

\item{data}{A data frame holding the time series data.}

\item{priors}{The priors for the parameters to fit. Each prior is
specified with a formula notation, for example, \code{beta ~
U(0, 1)} to specify that beta is uniformly distributed between
0 and 1. Use \code{c()} to provide more than one prior, for
example, \code{c(beta ~ U(0, 1), gamma ~ N(10, 1)}. Gamma
\code{G}, normal \code{N} and uniform \code{U} distributions
are supported.}

\item{npart}{An integer specifying the number of particles for the
bootstrap particle filter.}

\item{niter}{An integer specifying the number of iterations to run
the PMCMC.}

\item{fn}{FIXME.}

\item{...}{Further arguments to be passed to \code{fn}.}

\item{verbose}{prints diagnostic messages when \code{TRUE}. The
default is to retrieve the global option \code{verbose} and
use \code{FALSE} if it is not set.}
}
\description{
Particle Markov chain Monte Carlo (PMCMC) algorithm
}
\references{
\Andrieu2010
}
