% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SISe.R, R/SISe3.R
\docType{methods}
\name{prevalence}
\alias{prevalence}
\alias{prevalence,SISe-method}
\alias{prevalence,SISe3-method}
\title{Prevalence}
\usage{
prevalence(model, ...)

\S4method{prevalence}{SISe}(model, wnp = FALSE, i = NULL, by = 1, ...)

\S4method{prevalence}{SISe3}(model, age = 1:3, wnp = FALSE, i = NULL,
  by = 1, ...)
}
\arguments{
\item{model}{The \code{model} to calculated the prevalence from}

\item{...}{Additional arguments affecting the measure}

\item{wnp}{Determine within-node prevalence. Default is FALSE.}

\item{i}{Indices specifying the nodes to include in the
calculation of the prevalence. If \code{wnp = TRUE}, then
specifying which nodes to extract prevalence for. Default is NULL,
which includes all nodes.}

\item{by}{The number to increment the sequence of time points
starting from 1. Default is 1, which gives the prevalence at every
time point.}

\item{age}{For models with age categories, the age category to
include in the calculation. Default is that all age categories are
included.}
}
\description{
Calculate the proportion infected individuals
}
\keyword{methods}

