% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SIR.R, R/SISe.R, R/SISe3.R,
%   R/SISe3_sp.R, R/SISe_sp.R
\docType{methods}
\name{prevalence}
\alias{prevalence}
\alias{prevalence,SIR-method}
\alias{prevalence,SISe-method}
\alias{prevalence,SISe3-method}
\alias{prevalence,SISe3_sp-method}
\alias{prevalence,SISe_sp-method}
\title{Prevalence}
\usage{
prevalence(model, ...)

\S4method{prevalence}{SIR}(model, wnp = FALSE, i = NULL, ...)

\S4method{prevalence}{SISe}(model, wnp = FALSE, i = NULL, ...)

\S4method{prevalence}{SISe3}(model, age = 1:3, wnp = FALSE, i = NULL, ...)

\S4method{prevalence}{SISe3_sp}(model, age = 1:3, wnp = FALSE, i = NULL,
  ...)

\S4method{prevalence}{SISe_sp}(model, wnp = FALSE, i = NULL, ...)
}
\arguments{
\item{model}{The \code{model} to calculated the prevalence from}

\item{...}{Additional arguments affecting the measure}

\item{wnp}{Determine within-node prevalence. Default is FALSE.}

\item{i}{Indices specifying the nodes to include in the
calculation of the prevalence. If \code{wnp = TRUE}, then
specifying which nodes to extract prevalence for. Default is NULL,
which includes all nodes.}

\item{age}{For models with age categories, the age category to
include in the calculation. Default is that all age categories are
included.}
}
\description{
Calculate the proportion infected individuals
}
\keyword{methods}
