% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SISe.R
\docType{methods}
\name{prevalence}
\alias{prevalence}
\alias{prevalence,SISe-method}
\title{Prevalence}
\usage{
prevalence(model, ...)

\S4method{prevalence}{SISe}(model, ...)
}
\arguments{
\item{model}{The \code{model} to calculated the prevalence from}

\item{...}{Additional arguments affecting the measure}
}
\description{
Calculate the proportion infected individuals
}
\keyword{methods}

