% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_model.R
\name{prevalence}
\alias{prevalence}
\title{Prevalence}
\usage{
prevalence(model, numerator, denominator, type = c("pop", "bnp", "wnp"),
  i = NULL)
}
\arguments{
\item{model}{The \code{model} to calculate the prevalence from.}

\item{numerator}{The compartments in the population with a disease
or a condition.}

\item{denominator}{The compartments that define the entire
population of interest.}

\item{type}{The type of prevalence measure to calculate:
\code{'pop'} (default) calcalates the proportion of the
individuals in the population that have disease (model
specific) at each time point in \code{tspan}, \code{'bnp'}
calculates the between-node prevalence, and \code{'wnp'}
calculates the within-node prevalence.}

\item{i}{Indices specifying the nodes to include in the
calculation of the prevalence. Default is \code{NULL}, which
includes all nodes.}
}
\value{
Vector when type equals \code{'pop'} or \code{'bnp'} but
    matrix when type equals \code{'wnp'}.
}
\description{
Calculate the proportion of individuals with disease, or the
proportion of nodes with individuals with disease, or the
proportion of individuals with disease in each node.
}
