% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_model.R
\name{prevalence}
\alias{prevalence}
\title{Calculate the prevalence of disease found in a trajectory}
\usage{
prevalence(model = NULL, cases = NULL, pop = NULL, type = c("pop",
  "nop", "wnp"), i = NULL, as.is = FALSE)
}
\arguments{
\item{model}{The \code{model} to calculate the prevalence from.}

\item{cases}{The compartments in the population with a disease or
a condition.}

\item{pop}{The compartments that define the entire population of
interest.}

\item{type}{The type of prevalence measure to calculate:
\code{'pop'} (default) calcalates the proportion of the
individuals in the population that have disease (model
specific) at each time point in \code{tspan}, \code{'nop'}
calculates the node prevalence, and \code{'wnp'} calculates
the within-node prevalence.}

\item{i}{Indices specifying the nodes to include in the
calculation of the prevalence. Default is \code{NULL}, which
includes all nodes.}

\item{as.is}{The default (\code{as.is = FALSE}) is to generate a
\code{data.frame} with one row per time-step with the
prevalence. Using \code{as.is = TRUE} returns the result as a
matrix, which is the internal format.}
}
\value{
Vector when type equals \code{'pop'} or \code{'nop'} but
    matrix when type equals \code{'wnp'}.
}
\description{
Calculate the proportion of individuals with disease, or the
proportion of nodes with individuals with disease, or the
proportion of individuals with disease in each node.
}
