% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/run.r
\docType{methods}
\name{run}
\alias{run}
\alias{run,SISe-method}
\alias{run,SISe3-method}
\title{Run siminf stochastic simulation algorithms}
\usage{
run(model, threads = NULL, seed = NULL)

\S4method{run}{SISe3}(model, threads = NULL, seed = NULL)

\S4method{run}{SISe}(model, threads = NULL, seed = NULL)
}
\arguments{
\item{model}{The siminf model to run.}

\item{threads}{Number of threads. Default is NULL, i.e. to use the
number of available processors.  called.}

\item{seed}{Random number seed.}
}
\value{
\code{siminf_model} with result from simulation.
}
\description{
Run siminf stochastic simulation algorithms
}

