% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/run.r
\docType{methods}
\name{run}
\alias{run}
\alias{run,SISe-method}
\alias{run,SISe3-method}
\title{Run siminf stochastic simulation algorithms}
\usage{
run(model, threads = 1, verbose = 0, seed = NULL)

\S4method{run}{SISe3}(model, threads = 1, verbose = 0, seed = NULL)

\S4method{run}{SISe}(model, threads = 1, verbose = 0, seed = NULL)
}
\arguments{
\item{model}{The siminf model to run.}

\item{threads}{Number of threads. Default is 1.}

\item{verbose}{Level of siminf feeedback during simulation. Silent
if 0, progress if 1, progress and number of transition
events if 2. Default is 0.}

\item{seed}{Random number seed.}
}
\value{
\code{siminf_model} with result from simulation.
}
\description{
Run siminf stochastic simulation algorithms
}

