% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/run.r
\docType{methods}
\name{run}
\alias{run}
\alias{run,SISe-method}
\alias{run,SISe3-method}
\title{Run siminf stochastic simulation algorithms}
\usage{
run(model, verbose = 0, seed = NULL, threads = 1, strategy = c("single",
  "omp", "sg"))

\S4method{run}{SISe3}(model, verbose = 0, seed = NULL, threads = 1,
  strategy = c("single", "omp", "sg"))

\S4method{run}{SISe}(model, verbose = 0, seed = NULL, threads = 1,
  strategy = c("single", "omp", "sg"))
}
\arguments{
\item{model}{The siminf model to run.}

\item{verbose}{Level of siminf feeedback during simulation. Silent
if 0, progress if 1, progress and number of transition
events if 2. Default is 0.}

\item{seed}{Random number seed.}

\item{threads}{Number of threads. Default is 1.}

\item{strategy}{The parallelization strategy. Default is 'single'.}
}
\value{
\code{siminf_model} with result from simulation.
}
\description{
Run siminf stochastic simulation algorithms
}

