% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R, R/abc.R
\name{run}
\alias{run}
\alias{run,SimInf_model-method}
\alias{run,SEIR-method}
\alias{run,SIR-method}
\alias{run,SIS-method}
\alias{run,SISe-method}
\alias{run,SISe3-method}
\alias{run,SISe3_sp-method}
\alias{run,SISe_sp-method}
\alias{run,SimInf_abc-method}
\title{Run a SimInf model simulation}
\usage{
run(model, ...)

\S4method{run}{SimInf_model}(model, ...)

\S4method{run}{SEIR}(model, ...)

\S4method{run}{SIR}(model, ...)

\S4method{run}{SIS}(model, ...)

\S4method{run}{SISe}(model, ...)

\S4method{run}{SISe3}(model, ...)

\S4method{run}{SISe3_sp}(model, ...)

\S4method{run}{SISe_sp}(model, ...)

\S4method{run}{SimInf_abc}(model, ...)
}
\arguments{
\item{model}{The SimInf model to run.}

\item{...}{Optional arguments that affect the simulation:
\describe{
  \item{seed}{Numeric or integer specifying the random seed
  for the simulation. If not provided, a seed is randomly
  sampled from the current R RNG state.}
}}
}
\value{
The model object with a single stochastic trajectory
    attached.
}
\description{
Simulate a \code{\link{trajectory}} from a SimInf model. The
function compiles and loads model-specific C code (if not already
compiled), initializes the solver, and advances the simulation in
continuous time from the first to the last time point in
\code{tspan}. The state of the system is recorded at each time
point specified in \code{tspan}. Sparse output can be requested
with \code{\link{punchcard<-}} to store only selected time points
or compartments.
}
\examples{
## For reproducibility, specify the number of threads to use.
set_num_threads(1)

## Create an 'SIR' model with 10 nodes.
model <- SIR(
  u0 = data.frame(
    S = rep(99, 10),
    I = rep(1, 10),
    R = rep(0, 10)
  ),
  tspan = 1:100,
  beta = 0.16,
  gamma = 0.077
)

## Run the model. For reproducibility, specify the seed.
result <- run(model, seed = 22)

## Plot the distribution of susceptible, infected and recovered
## individuals.
plot(result)
}
\references{
\Widgren2019

\Bauer2016

\Bauer2015
}
