% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduled_events.R
\name{scheduled_events}
\alias{scheduled_events}
\title{Create a \code{"\linkS4class{scheduled_events}"} object}
\usage{
scheduled_events(E = NULL, N = NULL, events = NULL)
}
\arguments{
\item{E}{Sparse matrix of object class \code{"\linkS4class{dgCMatrix}"}
that selects the states to include for sampling in an event.}

\item{N}{Sparse matrix of object class \code{"\linkS4class{dgCMatrix}"}
that determines how to shift the states in an
\code{INTERNAL_TRANSFER_EVENT}.}

\item{events}{A \code{data.frame} with events.}
}
\value{
S4 class \code{scheduled_events}
}
\description{
The argument events must be a \code{data.frame} with the following
columns:
\describe{
  \item{event}{
    The event type. The following four type of events exists;
    \code{EXIT_EVENT = 0}, \code{ENTER_EVENT = 1},
    \code{INTERNAL_TRANSFER_EVENT = 2}, and
    \code{EXTERNAL_TRANSFER_EVENT = 3}.
  }
  \item{time}{
    Time for the event. Integer vector of length \code{len}.
  }
  \item{node}{
    The node that the event operates on. Also the source node for
    an \emph{external transfer} event.
    1 <= \code{node[i]} <= Number of nodes.
  }
  \item{dest}{
    The destination node for an \emph{external transfer} event.
    Should be \code{0} for the other event types. dest[i] >= 0.
  }
  \item{n}{
    The number of individuals affected by the event. n[i] >= 0.
  }
  \item{proportion}{
    If \code{n[i]} equals zero, the number of individuals affected by
    \code{event[i]} is calculated by summing the number of individuls
    with \code{proportion[i]}. 0 <= proportion[i] <= 1.
  }
  \item{select}{
    The column \code{j} in the event matrix \code{E} that
    determines the compartments that the event operates on.
  }
  \item{shift}{
    The column \code{k} in the shift matrix \code{N} that
    determines how individuals in \emph{internal transfer} and
    \emph{external transfer} events are shifted to enter another
    compartment.  Should be \code{0} for the other event types.
  }
}
}

