% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmp.R
\name{set_num_threads}
\alias{set_num_threads}
\title{Specify the number of threads to use}
\usage{
set_num_threads(threads = NULL)
}
\arguments{
\item{threads}{integer with maximum number of threads to use in
functions that are parallelized with OpenMP (if
available). Default is NULL, i.e. to use all available
processors and then check for limits in the environment
varibles.}
}
\value{
Integer with the maximum number of threads that will be
    used.
}
\description{
If OpenMP is available for parallelization, SimInf uses the
'omp_get_num_procs()' function and the environmental variables
'OMP_THREAD_LIMIT', 'OMP_NUM_THREADS', and 'SIMINF_NUM_THREADS' to
determine the maximum number of threads to use in functions that
are parallelized. Additionally, the maximum number of threads can
be controlled by the 'threads' argument.
}
