% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{show,SimInf_abc-method}
\alias{show,SimInf_abc-method}
\title{Print summary of a \code{SimInf_abc} object}
\usage{
\S4method{show}{SimInf_abc}(object)
}
\arguments{
\item{object}{The \code{SimInf_abc} object.}
}
\value{
\code{invisible(object)}.
}
\description{
Print summary of a \code{SimInf_abc} object
}
