% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_events.R
\docType{methods}
\name{show,SimInf_events-method}
\alias{show,SimInf_events-method}
\alias{show,SimInf_events-methods}
\title{Brief summary of \code{SimInf_events}}
\usage{
\S4method{show}{SimInf_events}(object)
}
\arguments{
\item{object}{The SimInf_events \code{object}}
}
\value{
None (invisible 'NULL').
}
\description{
Shows the number of scheduled events.
}
\keyword{methods}
