% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_model.R
\docType{methods}
\name{show,siminf_model-method}
\alias{show,siminf_model-method}
\alias{show,siminf_model-methods}
\title{Brief summary of \code{siminf_model}}
\usage{
\S4method{show}{siminf_model}(object)
}
\arguments{
\item{object}{The siminf_model \code{object}}
}
\value{
None (invisible 'NULL').
}
\description{
Brief summary of \code{siminf_model}
}
\examples{
## Create a 'SISe' demo model with 1 node and initialize
## it to run over 1000 days.
model <- demo_model(nodes = 1, days = 1000, model = "SISe")

## Brief summary of the model
model

## Run the model and save the result
result <- run(model)

## Brief summary of the result.
result
}
\keyword{methods}
