% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siminf_model.R
\docType{methods}
\name{show,siminf_model-method}
\alias{show,siminf_model-method}
\alias{show,siminf_model-methods}
\title{Brief summary of \code{siminf_model}}
\usage{
\S4method{show}{siminf_model}(object)
}
\arguments{
\item{object}{The siminf_model \code{object}}
}
\value{
None (invisible 'NULL').
}
\description{
Brief summary of \code{siminf_model}
}
\keyword{methods}

