% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_model.R
\docType{methods}
\name{show,SimInf_model-method}
\alias{show,SimInf_model-method}
\alias{show,SimInf_model-methods}
\title{Brief summary of \code{SimInf_model}}
\usage{
\S4method{show}{SimInf_model}(object)
}
\arguments{
\item{object}{The SimInf_model \code{object}}
}
\value{
None (invisible 'NULL').
}
\description{
Brief summary of \code{SimInf_model}
}
\examples{
## Create an 'SIR' model with 10 nodes and initialise
## it to run over 100 days.
model <- SIR(u0 = data.frame(S = rep(99, 10),
                             I = rep(1, 10),
                             R = rep(0, 10)),
             tspan = 1:100,
             beta = 0.16,
             gamma = 0.077)

## Brief summary of the model
model

## Run the model and save the result
result <- run(model, threads = 1, seed = 1)

## Brief summary of the result. Note that 'U' and 'V' are
## non-empty after running the model.
result
}
\keyword{methods}
