% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf.R
\docType{package}
\name{SimInf}
\alias{SimInf}
\alias{SimInf-package}
\title{A Framework for Stochastic Disease Spread Simulations}
\description{
AEM. v.0.1 
A Framework for Stochastic Disease Spread Simulations
}
