% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\docType{class}
\name{SimInf_abc_smc-class}
\alias{SimInf_abc_smc-class}
\title{Class \code{"SimInf_abc_smc"}}
\description{
Class \code{"SimInf_abc_smc"}
}
\section{Slots}{

\describe{
\item{\code{model}}{The \code{SimInf_model} object to estimate parameters
in.}

\item{\code{priors}}{FIXME.}

\item{\code{target}}{Character vector (\code{gdata} or \code{ldata}) that
determines if the ABC-SMC method estimates parameters in
\code{model@gdata} or in \code{model@ldata}.}

\item{\code{pars}}{Index to the parameters in \code{target}.}

\item{\code{npart}}{The number of particles in each generation.}

\item{\code{nprop}}{An integer vector with the number of simulated
proposals in each generation.}

\item{\code{fn}}{FIXME.}

\item{\code{x}}{A list where each item is a \code{matrix} with the
accepted particles in each generation. Each column is one
particle.}

\item{\code{w}}{A list where each item is a vector with the weights for
the particles \code{x} in the corresponding generation.}

\item{\code{ess}}{A numeric vector with the effective sample size (ESS) in
each generation. Effective sample size is computed as
\deqn{\left(\sum_{i=1}^N\!(w_{g}^{(i)})^2\right)^{-1},}{1 /
(sum(w_ig^2)),} where \eqn{w_{g}^{(i)}}{w_ig} is the
normalized weight of particle \eqn{i} in generation \eqn{g}.}
}}

