% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_smc.R
\docType{class}
\name{SimInf_abc_smc-class}
\alias{SimInf_abc_smc-class}
\title{Class \code{"SimInf_abc_smc"}}
\description{
Class \code{"SimInf_abc_smc"}
}
\section{Slots}{

\describe{
  \item{model}{
    FIXME.
  }
  \item{priors}{
    FIXME.
  }
  \item{target}{
    FIXME.
  }

  \item{i}{
    index to the parameters in \code{target}.
  }

  \item{npart}{
    The number of particles in each generation.
  }
  \item{nprop}{
    An integer vector with the number of simulated proposals in
    each generation.
  }
  \item{fn}{
    FIXME.
  }
  \item{x}{
    A list where each item is a \code{matrix} with the accepted
    particles in each generation. Each column is one particle.
  }
  \item{w}{
    A list where each item is a vector with the weights for the
    particles \code{x} in the corresponding generation.
  }
  \item{ess}{
    A numeric vector with the effective sample size (ESS) in each
    generation. Effective sample size is computed as
    \deqn{\left(\sum_{i=1}^N\!(w_{g}^{(i)})^2\right)^{-1},}{1/(sum(w_ig^2)),}
    where \eqn{w_{g}^{(i)}}{w_ig} is the normalized weight of
    particle \eqn{i} in generation \eqn{g}.
  }
}
}

