% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_individual_events.R
\docType{class}
\name{SimInf_individual_events-class}
\alias{SimInf_individual_events-class}
\title{Class \code{"SimInf_individual_events"}}
\description{
Class \code{"SimInf_individual_events"}
}
\section{Slots}{

\describe{
\item{\code{id}}{an integer or character identifier of the individual.}

\item{\code{event}}{four event types are supported: \emph{exit},
\emph{enter}, \emph{internal transfer}, and \emph{external
transfer}.  When assigning the events, they can either be
coded as a numerical value or a character string: \emph{exit;}
\code{0} or \code{'exit'}, \emph{enter;} \code{1} or
\code{'enter'}, \emph{internal transfer;} \code{2} or
\code{'intTrans'}, and \emph{external transfer;} \code{3} or
\code{'extTrans'}.}

\item{\code{time}}{an integer, character, or date (of class \code{Date})
for when the event occured. If it's a character it must be
able to coerce to \code{Date}.}

\item{\code{node}}{an integer or character identifier of the source node.}

\item{\code{dest}}{an integer or character identifier of the destination
node.}
}}

