% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/siminf_model.R
\docType{class}
\name{siminf_model-class}
\alias{siminf_model-class}
\title{Class \code{"siminf_model"}}
\description{
Class to handle the siminf data model
}
\section{Slots}{

\describe{
  \item{G}{
    Sparse matrix (\eqn{Nt \times Nt}) of object class
    \code{"\linkS4class{dgCMatrix}"}.  A non-zeros entry in element
    \code{i} of column \code{j} indicates that propensity \code{i}
    needs to be recalculated if the transition \code{j} occurs.
  }
  \item{N}{
    Sparse matrix (\eqn{Nc \times Nt}) of object class
    \code{"\linkS4class{dgCMatrix}"}. Each column corresponds
    to a transition, and execution of transition \code{j} amounts to
    adding the \code{j}th column to the state vector.
  }
  \item{U}{
    The result matrix ((Nn * Nc) X length(tspan)). U(:,j) contains
    the state of the system at tspan(j).
  }
  \item{Nn}{
    Number of nodes.
  }
  \item{data}{
    A data vector passed as an argument to the propensities.
  }
  \item{sd}{
    Integer vector of length Nn. Each node can be assigned to a sub-domain.
  }
  \item{tspan}{
    A vector of increasing time points where the state of each node is
    to be returned.
  }
  \item{u0}{
     Initial state vector u0. Integer (Nc X Nn). Gives the initial
     number of individuals in each compartment in every node.
  }
  \item{events}{
    External events \code{"\linkS4class{external_events}"}
  }
}
}
\keyword{classes}
\keyword{methods}

