% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siminf_model.R
\docType{class}
\name{siminf_model-class}
\alias{siminf_model-class}
\title{Class \code{"siminf_model"}}
\description{
Class to handle the siminf data model
}
\section{Slots}{

\describe{
  \item{G}{
    Sparse matrix (\eqn{Nt \times Nt}) of object class
    \code{"\linkS4class{dgCMatrix}"}.  A non-zeros entry in element
    \code{i} of column \code{j} indicates that propensity \code{i}
    needs to be recalculated if the transition \code{j} occurs.
  }
  \item{S}{
    Sparse matrix (\eqn{Nc \times Nt}) of object class
    \code{"\linkS4class{dgCMatrix}"}. Each column corresponds
    to a transition, and execution of transition \code{j} amounts to
    adding the \code{j}th column to the state vector.
  }
  \item{U}{
    The compartment result matrix ((Nn * Nc) X length(tspan)).
    U(:,j) contains the state of the compartments at tspan(j).
  }
  \item{ldata}{
    A matrix with local data for nodes. The column ldata[, j] contains
    the local data vector for node #j. The local data vector is passed
    as an argument to the propensities and the post time step function.
  }
  \item{gdata}{
    A vector with global data for all nodes. The global data vector
    is passed as an argument to the propensities and the post time
    step function.
  }
  \item{sd}{
    Integer vector of length Nn. Each node can be assigned to a sub-domain.
  }
  \item{tspan}{
    A vector of increasing time points where the state of each node is
    to be returned.
  }
  \item{u0}{
     Initial compartment state vector u0. Integer (Nc X Nn). Gives the initial
     number of individuals in each compartment in every node.
  }
  \item{V}{
    The model parameter result matrix for parameters that are modelled
    ((Nn * Number of paramters) X length(tspan)).
    V(:,j) contains the state of the model parameters at tspan(j).
  }
  \item{v0}{
     Initial model state vector v0. Double (Number of parameters X Nn).
     Gives the initial value for model paramenters that are updated
     during the simulation in every node.
  }
  \item{events}{
    External events \code{"\linkS4class{scheduled_events}"}
  }
}
}
\keyword{classes}
\keyword{methods}

