% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_model.R
\name{SimInf_model}
\alias{SimInf_model}
\title{Create a \code{SimInf_model}}
\usage{
SimInf_model(
  G,
  S,
  tspan,
  events = NULL,
  ldata = NULL,
  gdata = NULL,
  U = NULL,
  u0 = NULL,
  v0 = NULL,
  V = NULL,
  E = NULL,
  N = NULL,
  C_code = NULL
)
}
\arguments{
\item{G}{Dependency graph that indicates the transition rates that
need to be updated after a given state transition has occured.
A non-zero entry in element \code{G[i, i]} indicates that
transition rate \code{i} needs to be recalculated if the state
transition \code{j} occurs. Sparse matrix (\eqn{Nt \times Nt})
of object class
\code{\link[Matrix:dgCMatrix-class]{dgCMatrix}}.}

\item{S}{Each column corresponds to a transition, and execution of
state transition \code{j} amounts to adding the \code{S[, j]}
to the state vector of the node where the state transition
occurred.  Sparse matrix (\eqn{Nc \times Nt}) of object class
\code{\link[Matrix:dgCMatrix-class]{dgCMatrix}}.}

\item{tspan}{A vector (length >= 1) of increasing time points
where the state of each node is to be returned. Can be either
an \code{integer} or a \code{Date} vector. A \code{Date}
vector is coerced to a numeric vector as days, where
\code{tspan[1]} becomes the day of the year of the first year
of \code{tspan}. The dates are added as names to the numeric
vector.}

\item{events}{A \code{data.frame} with the scheduled events.}

\item{ldata}{local data for the nodes. Can either be specified as
a \code{data.frame} with one row per node. Or as a matrix
where each column \code{ldata[, j]} contains the local data
vector for the node \code{j}. The local data vector is passed
as an argument to the transition rate functions and the post
time step function.}

\item{gdata}{A numeric vector with global data that is common to
all nodes. The global data vector is passed as an argument to
the transition rate functions and the post time step function.}

\item{U}{The result matrix with the number of individuals in each
disease state in every node (\eqn{N_n N_c \times}
\code{length(tspan)}).  \code{U[, j]} contains the number of
individuals in each disease state at
\code{tspan[j]}. \code{U[1:Nc, j]} contains the state of node
\code{1} at \code{tspan[j]}. \code{U[(Nc + 1):(2 * Nc), j]}
contains the state of node \code{2} at \code{tspan[j]} etc.}

\item{u0}{A \code{data.frame} with the initial state in each node,
i.e., the number of individuals in each compartment in each
node when the simulation starts (see \sQuote{Details}). The
parameter \code{u0} can also be an object that can be coerced
to a \code{data.frame}, e.g., a named numeric vector will be
coerced to a one row \code{data.frame}.}

\item{v0}{The initial continuous state vector in every node.
(\code{dim(ldata)[1]} \eqn{\times N_N}). The continuous state
vector is updated by the specific model during the simulation
in the post time step function.}

\item{V}{The result matrix for the real-valued continous
compartment state (\eqn{N_n}\code{dim(ldata)[1]} \eqn{\times}
\code{length(tspan)}).  \code{V[, j]} contains the real-valued
state of the system at \code{tspan[j]}.}

\item{E}{A matrix to handle scheduled events, see
\code{\linkS4class{SimInf_events}}.  Each row in the matrix
corresponds to one compartment in the model. The non-zero
entries in a column indicate the compartments to include in an
event. For the \emph{exit}, \emph{internal transfer} and
\emph{external transfer} events, the values in \code{E[,
select]} are used as weights when sampling individuals without
replacement, with probability proportional to the weight. For
the \emph{enter} event, the values in \code{E[, select]} are
used as weights when determining which compartment to add
individuals to. If the column \code{E[, select]} contains
several non-zero entries, the compartment is sampled with
probability proportional to the weight in \code{E[, select]}.
The select matrix \code{E} can either be specified as a
\code{matrix}, or as a \code{data.frame}.  When \code{E} is
specified as a \code{data.frame}, it must have one column
named \code{compartment} that defines which compartment is
referred to, and one column \code{select} that defines the
column in \code{E}.  In addition, the \code{data.frame} can
contain an optional column named \code{value} with the value
in \code{E}.  When the \code{value} column is missing,
\code{1} is used as the default value.}

\item{N}{A matrix to handle scheduled events, see
\code{\linkS4class{SimInf_events}}. Each row in the matrix
corresponds to one compartment in the model. The values in a
column define how to move sampled individuals before adding
them to the destination. Let \code{q <- shift}, then each
non-zero entry in \code{N[, q]} defines the number of rows to
move sampled individuals from that compartment i.e., sampled
individuals from compartment \code{p} are moved to compartment
\code{N[p, q] + p}, where \code{1 <= N[p, q] + p <=
N_compartments}. This matrix is used for \emph{enter},
\emph{internal transfer} and \emph{external transfer} events.
The shift matrix \code{N} can either be specified as a
\code{matrix}, or as a \code{data.frame}.  When \code{N} is
specified as a \code{data.frame}, it must have one column
named \code{compartment} that defines which compartment is
referred to, and one column \code{shift} that defines the
column in \code{N}.  In addition, the \code{data.frame} must
contain a column named \code{value} with the integer value in
\code{N}.}

\item{C_code}{Character vector with optional model C code. If
non-empty, the C code is written to a temporary C-file when
the \code{run} method is called.  The temporary C-file is
compiled and the resulting DLL is dynamically loaded. The DLL
is unloaded and the temporary files are removed after running
the model.}
}
\value{
\linkS4class{SimInf_model}
}
\description{
Create a \code{SimInf_model}
}
