% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siminf_model.R
\name{siminf_model}
\alias{siminf_model}
\title{Create a siminf model}
\usage{
siminf_model(G, N, tspan, events = NULL, sd = NULL, ldata = NULL,
  gdata = NULL, U = NULL, Nn = NULL, u0 = NULL, init = NULL,
  v0 = NULL, V = NULL, E = NULL, S = NULL)
}
\arguments{
\item{G}{Sparse matrix (\eqn{Nt \times Nt}) of object class
\code{"\linkS4class{dgCMatrix}"}.  A non-zeros entry in element
\code{i} of column \code{j} indicates that propensity \code{i}
needs to be recalculated if the transition \code{j} occurs.}

\item{N}{Sparse matrix (\eqn{Nc \times Nt}) of object class
\code{"\linkS4class{dgCMatrix}"}. Each column corresponds to a
transition, and execution of transition \code{j} amounts to adding
the \code{j}th column to the state vector.}

\item{tspan}{A vector of increasing time points where the state of
each node is to be returned.}

\item{events}{A \code{data.frame} with the scheduled events.}

\item{sd}{Integer vector of length Nn. Each node can be assigned
to a sub-domain.}

\item{ldata}{A matrix with local data for nodes. The column
ldata[, j] contains the local data vector for node #j. The local
data vector is passed as an argument to the propensities and the
post time step function.}

\item{gdata}{A vector with global data for all nodes. The global
data vector is passed as an argument to the propensities and the
post time step function.}

\item{U}{The result matrix ((Nn * Nc) X length(tspan)). U(:,j)
contains the state of the system at tspan(j).}

\item{Nn}{Number of nodes.}

\item{u0}{Initial state vector u0. Integer (Nc X Nn). Gives the
initial number of individuals in each compartment in every node.}

\item{init}{A \code{data.frame} with the initial number of
individuals in each compartment in every node.}

\item{v0}{Initial model state vector v0. Double (Number of
parameters X Nn).  Gives the initial value for model paramenters
that are updated during the simulation in every node.}

\item{V}{The model parameter result matrix for the parameters that
are modelled ((Nn * Number of paramters) X length(tspan)).  V(:,j)
contains the state of the model parameters at tspan(j).}

\item{E}{Sparse matrix to handle external events, see
\code{\linkS4class{external_events}}.}

\item{S}{Sparse matrix to handle external events, see
\code{\linkS4class{external_events}}.}
}
\value{
\linkS4class{siminf_model}
}
\description{
Create a siminf model
}

