% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfilter.R
\docType{class}
\name{SimInf_pfilter-class}
\alias{SimInf_pfilter-class}
\title{Class \code{"SimInf_pfilter"}}
\description{
Class \code{"SimInf_pfilter"}
}
\section{Slots}{

\describe{
\item{\code{model}}{The \code{SimInf_model} object to simulate data from.}

\item{\code{npart}}{An integer with the number of particles that was used
at each timestep.}

\item{\code{tspan}}{A two-column matrix where each row specifies the tspan
to use when running the model from time[i] to time[i+1]. The
first column is \code{NA} if the interval is one time-unit.}

\item{\code{loglik}}{The estimated log likelihood.}
}}

