% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfilter.R
\docType{class}
\name{SimInf_pfilter-class}
\alias{SimInf_pfilter-class}
\title{Class \code{"SimInf_pfilter"}}
\description{
Class \code{"SimInf_pfilter"}
}
\section{Slots}{

\describe{
\item{\code{model}}{A \code{SimInf_model} object with one filtered
trajectory attached.}

\item{\code{n_particles}}{An integer with the number of particles that was
used at each timestep.}

\item{\code{loglik}}{The estimated log likelihood.}

\item{\code{ess}}{A numeric vector with the effective sample size (ESS).
The effective sample size is computed as
\deqn{\left(\sum_{i=1}^N\!(w_{t}^{i})^2\right)^{-1},}{1 /
(sum(w_it^2)),} where \eqn{w_{t}^{i}}{w_it} is the normalized
weight of particle \eqn{i} at time \eqn{t}.}
}}

