% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{SimInf_raster_model-class}
\alias{SimInf_raster_model-class}
\title{Class \code{"SimInf_raster_model"}}
\description{
Class to handle data for the \code{SimInf_raster_model}.  It is a
model where the nodes are not fixed at one position but can move
between cells in a raster.
}
\section{Slots}{

\describe{
\item{\code{cell_S}}{FIXME.}

\item{\code{cell_U}}{FIXME.}

\item{\code{cell_U_sparse}}{FIXME.}
}}

