% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_raster_model.R
\name{SimInf_raster_model}
\alias{SimInf_raster_model}
\title{Create a \code{SimInf_raster_model}}
\usage{
SimInf_raster_model(
  G,
  S,
  cell_S,
  tspan,
  ldata = NULL,
  gdata = NULL,
  u0 = NULL,
  v0 = NULL,
  C_code = NULL
)
}
\arguments{
\item{G}{Dependency graph that indicates the transition rates that
need to be updated after a given state transition has occured.
A non-zero entry in element \code{G[i, i]} indicates that
transition rate \code{i} needs to be recalculated if the state
transition \code{j} occurs. Sparse matrix (\eqn{Nt \times Nt})
of object class
\code{\link[Matrix:dgCMatrix-class]{dgCMatrix}}.}

\item{S}{Each column corresponds to a transition, and execution of
state transition \code{j} amounts to adding the \code{S[, j]}
to the state vector of the node where the state transition
occurred.  Sparse matrix (\eqn{Nc \times Nt}) of object class
\code{\link[Matrix:dgCMatrix-class]{dgCMatrix}}.}

\item{cell_S}{FIXME.}

\item{tspan}{A vector (length >= 1) of increasing time points
where the state of each node is to be returned. Can be either
an \code{integer} or a \code{Date} vector. A \code{Date}
vector is coerced to a numeric vector as days, where
\code{tspan[1]} becomes the day of the year of the first year
of \code{tspan}. The dates are added as names to the numeric
vector.}

\item{ldata}{local data for the nodes. Can either be specified as
a \code{data.frame} with one row per node. Or as a matrix
where each column \code{ldata[, j]} contains the local data
vector for the node \code{j}. The local data vector is passed
as an argument to the transition rate functions and the post
time step function.}

\item{gdata}{A numeric vector with global data that is common to
all nodes. The global data vector is passed as an argument to
the transition rate functions and the post time step function.}

\item{u0}{A \code{data.frame} with the initial state in each node,
i.e., the number of individuals in each compartment in each
node when the simulation starts (see \sQuote{Details}). The
parameter \code{u0} can also be an object that can be coerced
to a \code{data.frame}, e.g., a named numeric vector will be
coerced to a one row \code{data.frame}.}

\item{v0}{The initial continuous state vector in every node.
(\code{dim(ldata)[1]} \eqn{\times N_N}). The continuous state
vector is updated by the specific model during the simulation
in the post time step function.}

\item{C_code}{Character vector with optional model C code. If
non-empty, the C code is written to a temporary C-file when
the \code{run} method is called.  The temporary C-file is
compiled and the resulting DLL is dynamically loaded. The DLL
is unloaded and the temporary files are removed after running
the model.}
}
\description{
Create a \code{SimInf_raster_model} object.  It is a model where
the nodes are not fixed at one position but can move between cells
on a raster.
}
