% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SISe.R
\name{SISe}
\alias{SISe}
\title{Create a SISe model}
\usage{
SISe(init, tspan, events = NULL, phi = NULL, upsilon = NULL,
  gamma = NULL, alpha = NULL, beta_t1 = NULL, beta_t2 = NULL,
  beta_t3 = NULL, beta_t4 = NULL, end_t1 = NULL, end_t2 = NULL,
  end_t3 = NULL, end_t4 = NULL, epsilon = NULL)
}
\arguments{
\item{init}{A \code{data.frame} with the initial state in each
node, see details.}

\item{tspan}{An increasing sequence of points in time where the
state of the system is to be returned.}

\item{events}{a \code{data.frame} with the scheduled events, see
\code{\link{siminf_model}}.}

\item{phi}{A numeric vector with the initial environmental
infectious pressure in each node. Default NULL which gives 0 in
each node.}

\item{upsilon}{The response rate from susceptible to infected due
to the environmental infectious pressure}

\item{gamma}{The recover rate from infected to susceptible}

\item{alpha}{The shed rate}

\item{beta_t1}{The decay of the environmental infectious pressure
in the first interval of the year.}

\item{beta_t2}{The decay of the environmental infectious pressure
in the second interval of the year.}

\item{beta_t3}{The decay of the environmental infectious pressure
in the third interval of the year.}

\item{beta_t4}{The decay of the environmental infectious pressure
in the fourth interval of the year.}

\item{epsilon}{The background infectious pressure}
}
\value{
\code{SISe}
}
\description{
Create a SISe model to be used by the simulation framework.
}
\details{
The argument init must be a \code{data.frame} with the following
columns:
\describe{
\item{id}{Node identifier that uniquely identifies each node. The
node identifiers must be zero-based, i.e. the first identifier
must be equal to zero.}
\item{S}{The number of sucsceptible}
\item{I}{The number of infected}
}
}

