% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SISe3.R
\name{SISe3}
\alias{SISe3}
\title{Create a SISe3 model}
\usage{
SISe3(init, tspan, events = NULL, phi = NULL, upsilon_1 = NULL,
  upsilon_2 = NULL, upsilon_3 = NULL, gamma_1 = NULL, gamma_2 = NULL,
  gamma_3 = NULL, alpha = NULL, beta_q1 = NULL, beta_q2 = NULL,
  beta_q3 = NULL, beta_q4 = NULL, epsilon = NULL)
}
\arguments{
\item{init}{A \code{data.frame} with the initial state in each
node, see details.}

\item{tspan}{An increasing sequence of points in time where the
state of the system is to be returned.}

\item{events}{a \code{data.frame} with the scheduled events, see
\code{\link{siminf_model}}.}

\item{phi}{A numeric vector with the initial environmental
infectious pressure in each node. Default NULL which gives 0 in
each node.}

\item{upsilon_1}{The response rate from susceptible to
infected due to the environmental infectious pressure in age
category 1}

\item{upsilon_2}{The response rate from susceptible to
infected due to the environmental infectious pressure in age
category 2}

\item{upsilon_3}{The response rate from susceptible to
infected due to the environmental infectious pressure in age
category 3}

\item{gamma_1}{The recover rate from infected to
susceptible for age category 1}

\item{gamma_2}{The recover rate from infected to
susceptible for age category 2}

\item{gamma_3}{The recover rate from infected to
susceptible for age category 3}

\item{alpha}{The shed rate}

\item{beta_q1}{The decay of the environmental infectious pressure
in the first quarter of the year.}

\item{beta_q2}{The decay of the environmental infectious pressure
in the second quarter of the year.}

\item{beta_q3}{The decay of the environmental infectious pressure
in the third quarter of the year.}

\item{beta_q4}{The decay of the environmental infectious pressure
in the fourth quarter of the year.}

\item{epsilon}{The background infectious pressure}
}
\value{
\code{SISe3}
}
\description{
Create a SISe3 model to be used by the simulation framework.
}
\details{
The argument init must be a \code{data.frame} with the following
columns:
\describe{
\item{id}{Node identifier that uniquely identifies each node. The
node identifiers must be zero-based, i.e. the first identifier
must be equal to zero.}
\item{S_1}{The number of sucsceptible in age category 1}
\item{I_1}{The number of infected in age category 1}
\item{S_2}{The number of sucsceptible in age category 2}
\item{I_2}{The number of infected in age category 2}
\item{S_3}{The number of sucsceptible in age category 3}
\item{I_3}{The number of infected in age category 3}
}
}

