% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SISe_sp.R
\name{SISe_sp}
\alias{SISe_sp}
\title{Create a \code{SISe_sp} model}
\usage{
SISe_sp(init, tspan, events = NULL, phi = NULL, upsilon = NULL,
  gamma = NULL, alpha = NULL, beta_t1 = NULL, beta_t2 = NULL,
  beta_t3 = NULL, beta_t4 = NULL, end_t1 = NULL, end_t2 = NULL,
  end_t3 = NULL, end_t4 = NULL, epsilon = NULL, coupling = NULL,
  distance = NULL)
}
\arguments{
\item{init}{A \code{data.frame} with the initial state in each
node, see details.}

\item{tspan}{An increasing sequence of points in time where the
state of the system is to be returned.}

\item{events}{a \code{data.frame} with the scheduled events, see
\code{\link{siminf_model}}.}

\item{phi}{A numeric vector with the initial environmental
infectious pressure in each node. Default NULL which gives 0
in each node.}

\item{upsilon}{Indirect transmission rate of the environmental
infectious pressure}

\item{gamma}{The recovery rate from infected to susceptible}

\item{alpha}{Shed rate from infected individuals}

\item{beta_t1}{The decay of the environmental infectious pressure
in the first interval of the year.}

\item{beta_t2}{The decay of the environmental infectious pressure
in the second interval of the year.}

\item{beta_t3}{The decay of the environmental infectious pressure
in the third interval of the year.}

\item{beta_t4}{The decay of the environmental infectious pressure
in the fourth interval of the year.}

\item{end_t1}{The non-inclusive day that ends interval 1.}

\item{end_t2}{The non-inclusive day that ends interval 2.}

\item{end_t3}{The non-inclusive day that ends interval 3.}

\item{end_t4}{The non-inclusive day that ends interval 4.}

\item{epsilon}{The background infectious pressure}

\item{coupling}{The coupling between neighboring nodes}

\item{distance}{The distance matrix between neighboring nodes}
}
\value{
\code{SISe_sp}
}
\description{
Create a \code{SISe_sp} model to be used by the simulation
framework.
}
\details{
The argument init must be a \code{data.frame} with one row for
each node with the following columns:
\describe{
\item{id}{Node identifier that uniquely identifies each node. The
node identifiers must be zero-based, i.e. the first identifier
must be equal to zero.}
\item{S}{The number of sucsceptible}
\item{I}{The number of infected}
}
}
\section{Beta}{

The time dependent beta is divided into four intervals of the year
\preformatted{
where 0 <= day < 365 and
0 <= end_t1 < end_t2 < end_t3 < end_t4 <= 365
or
0 <= end_t4 < end_t1 < end_t2 < end_t3 < 364

Case 1: end_t1 < end_t4
INTERVAL_1  INTERVAL_2       INTERVAL_3       INTERVAL_4       INTERVAL_1
[0, end_t1) [end_t1, end_t2) [end_t2, end_t3) [end_t3, end_t4) [end_t4, 365]

Case 2: end_t4 < end_t1
INTERVAL_4  INTERVAL_1       INTERVAL_2       INTERVAL_3       INTERVAL_4
[0, end_t4) [end_t4, end_t1) [end_t1, end_t2) [end_t2, end_t3) [end_t3, 365)
}
}

