% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{summary,SimInf_abc-method}
\alias{summary,SimInf_abc-method}
\title{Detailed summary of a \code{SimInf_abc} object}
\usage{
\S4method{summary}{SimInf_abc}(object, ...)
}
\arguments{
\item{object}{The \code{\linkS4class{SimInf_abc}} object to
summarize.}

\item{...}{Additional arguments affecting the summary produced.
Currently ignored.}
}
\value{
\code{NULL}, returned invisibly.
}
\description{
Display a detailed summary of an approximate Bayesian computation
(ABC) analysis, including the number of particles per generation,
the number of generations, and—for each generation—the acceptance
rate, effective sample size (ESS), and summary statistics
(minimum, first quartile, median, mean, third quartile, and
maximum) for each parameter in the posterior distribution.
}
\details{
Compared to \code{\link[=show,SimInf_abc-method]{show}}, the
summary additionally displays posterior statistics for all
generations, not just the final one.
}
\seealso{
\code{\link{abc}} for running an ABC analysis,
    \code{\link{continue_abc}} for continuing an existing ABC
    chain, \code{\linkS4class{SimInf_abc}} for the class
    definition, \code{\link[=show,SimInf_abc-method[{show}} for a
    brief summary.
}
