% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{summary,SimInf_abc-method}
\alias{summary,SimInf_abc-method}
\title{Detailed summary of a \code{SimInf_abc} object}
\usage{
\S4method{summary}{SimInf_abc}(object, ...)
}
\arguments{
\item{object}{The \code{SimInf_abc} object}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
None (invisible 'NULL').
}
\description{
Detailed summary of a \code{SimInf_abc} object
}
