% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_events.R
\docType{methods}
\name{summary,SimInf_events-method}
\alias{summary,SimInf_events-method}
\title{Summary of \code{SimInf_events}}
\usage{
\S4method{summary}{SimInf_events}(object, ...)
}
\arguments{
\item{object}{The \code{SimInf_events} object}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
None (invisible 'NULL').
}
\description{
Shows the number of scheduled events and the number of scheduled
events per event type.
}
