% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_individual_events.R
\name{summary,SimInf_individual_events-method}
\alias{summary,SimInf_individual_events-method}
\title{Detailed summary of a \code{SimInf_individual_events} object}
\usage{
\S4method{summary}{SimInf_individual_events}(object, ...)
}
\arguments{
\item{object}{The \code{SimInf_individual_events} object}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
None (invisible 'NULL').
}
\description{
Detailed summary of a \code{SimInf_individual_events} object
}
