% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siminf_model.R
\docType{methods}
\name{summary,siminf_model-method}
\alias{summary,siminf_model-method}
\alias{summary,siminf_model-methods}
\title{Summary of \code{siminf_model}}
\usage{
\S4method{summary}{siminf_model}(object, ...)
}
\arguments{
\item{object}{The \code{siminf_model} object}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
None (invisible 'NULL').
}
\description{
Summary of \code{siminf_model}
}
\keyword{methods}

