% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_model.R
\docType{methods}
\name{summary,SimInf_model-method}
\alias{summary,SimInf_model-method}
\alias{summary,SimInf_model-methods}
\title{Summary of \code{SimInf_model}}
\usage{
\S4method{summary}{SimInf_model}(object, ...)
}
\arguments{
\item{object}{The \code{SimInf_model} object}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
None (invisible 'NULL').
}
\description{
Summary of \code{SimInf_model}
}
\keyword{methods}
