% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{summary,SimInf_model-method}
\alias{summary,SimInf_model-method}
\title{Detailed summary of a \code{SimInf_model} object}
\usage{
\S4method{summary}{SimInf_model}(object, ...)
}
\arguments{
\item{object}{The \code{SimInf_model} object}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
None (invisible 'NULL').
}
\description{
Detailed summary of a \code{SimInf_model} object
}
