% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfilter.R
\name{summary,SimInf_pfilter-method}
\alias{summary,SimInf_pfilter-method}
\title{Print summary of a \code{SimInf_pfilter} object}
\usage{
\S4method{summary}{SimInf_pfilter}(object, ...)
}
\arguments{
\item{object}{The \code{SimInf_pfilter} object.}

\item{...}{Unused additional arguments.}
}
\value{
\code{invisible(NULL)}.
}
\description{
Print summary of a \code{SimInf_pfilter} object
}
