% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SISe.R, R/SISe3.R
\docType{methods}
\name{susceptible}
\alias{susceptible}
\alias{susceptible,SISe-method}
\alias{susceptible,SISe3-method}
\title{Susceptible}
\usage{
susceptible(model, ...)

\S4method{susceptible}{SISe}(model, i = NULL, by = 1, ...)

\S4method{susceptible}{SISe3}(model, age = 1:3, i = NULL, by = 1, ...)
}
\arguments{
\item{model}{The \code{model} to extract the susceptible from}

\item{...}{Additional arguments affecting the measure}

\item{i}{Indices specifying the nodes to include when extracting
the number of susceptible. Default is NULL, which includes all
nodes.}

\item{by}{The number to increment the sequence of time points
starting from 1. Default is 1, which gives the number of
susceptible at every time point.}

\item{age}{For models with age categories, the age category to
extract.}
}
\description{
Extracts the number of susceptible
}
\keyword{methods}

