% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u0.R
\name{u0}
\alias{u0}
\alias{u0,SimInf_model-method}
\title{Get the initial compartment state}
\usage{
u0(model, ...)

\S4method{u0}{SimInf_model}(model, ...)
}
\arguments{
\item{model}{The model to get the initial compartment state
\code{u0} from.}

\item{...}{Additional arguments. Currently not used.}
}
\value{
a \code{data.frame} with the initial compartment state.
}
\description{
Get the initial compartment state
}
\examples{
## Create an SIR model object.
model <- SIR(u0 = data.frame(S = 99, I = 1, R = 0),
             tspan = 1:100,
             beta = 0.16,
             gamma = 0.077)

## Get the initial compartment state.
u0(model)
}
