% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR.R
\name{u0_SEIR}
\alias{u0_SEIR}
\title{Example data to initialize the \code{SEIR} model}
\usage{
u0_SEIR()
}
\value{
A \code{data.frame}
}
\description{
Synthetic init data for 1600 nodes to demonstrate the \code{SEIR}
model.
}
\keyword{methods}

