% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR.R
\name{u0_SIR}
\alias{u0_SIR}
\title{Example data to initialize the \code{SIR} model}
\usage{
u0_SIR()
}
\value{
A \code{data.frame}
}
\description{
Synthetic init data for 1600 nodes to demonstrate the \code{SIR}
model.
}
\keyword{methods}

