% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SISe.R
\name{u0_SISe}
\alias{u0_SISe}
\title{Example data to initialize the \code{SISe} model}
\usage{
u0_SISe()
}
\value{
A \code{data.frame}
}
\description{
Synthetic init data for 1600 nodes to demonstrate the \code{SISe}
model.
}
