% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf.R
\docType{data}
\name{u0_SISe3}
\alias{u0_SISe3}
\title{Example data to initialize the \code{SISe3} model}
\format{A \code{data.frame}}
\usage{
data(u0_SISe3)
}
\description{
Synthetic init data for 1600 nodes to demonstrate the \code{SISe3}
model.
}
\keyword{dataset}
